### 2019-01-21 #279 @aaronwolen
osfr: R interface to OSF
Submitting Author: Aaron Wolen (@aaronwolen)
Repository: https://github.com/CenterForOpenScience/osfr
Version submitted: 0.2.1
Version accepted: TBD

---

```
Package: osfr
Title: R Interface to OSF
Version: 0.2.1
    person("Aaron", "Wolen",, "aaron@wolen.com", role = c("aut", "cre"),
    person("Chris","Hartgerink",, "chjh@protonmail.com", role = "aut"),
    person("Timothy", "York",, "timothypyork@gmail.com", role = "ctb"),
    person("Richards", "Brian",, "brian.g.richards@gmail.com", role = "ctb"),
    person("Courtney", "Soderberg",, "courtney@cos.io", role = "ctb"))
Description: Manage your projects and access open research materials and data on OSF.
    crul,
    stringi,
    fs
License: Apache License 2.0 | file LICENSE
URL:
    http://centerforopenscience.github.io/osfr,
    https://github.com/centerforopenscience/osfr
    lintr,
    covr
```

- Explain how the and why the package falls under these categories (briefly, 1-2 sentences):

  osfr provides a suite of functions for interacting with OSF (https://osf.io), a project management repository that supports researchers across their entire project lifecycle. Using osfr and OSF, researchers can create and manage projects, upload and share materials, or explore publicly accessible projects and download the associated files.

    Any R users who rely on OSF for managing their research or working with collaborators.

    No.
