### 2019-01-22 4f638f8
Correct error in osf_download() docs (fixes #85)
