### 2019-01-23 1f601ca
Indicate all PAT scopes are required for osfr development
