### 2019-01-23 e6db814
Add contributing guideliens and CoC to README
