+++ README.md
@@ -194,4 +194,11 @@ Big thanks to Rusty Speidel for designing our logo and [Tim
 Errington](https://github.com/timerrington) for his feedback during
 development.
 
+## Contributing
+
+Check out the [Contributing Guidelines](.github/CONTRIBUTING.md) to get
+started with osfr development and note that by contributing to this
+project, you agree to abide by the terms outlined in the [Contributor
+Code of Conduct](CODE_OF_CONDUCT.md).
+
 <!-- links -->
