### 2019-01-24 81cfe71
Restructure tests to handle different environments
Tests should now run when:
* OSF_SERVER is not "test"
* a PAT is not registered
