### 2019-01-24 d8918f2
Suppress osf_auth() message during tests
