### 2019-01-24 230cb65
Update NEWS
