### 2019-01-24 71adf4f
Integrate osf_mv() tests with new testing strategy
