+++ tests/testthat/test-files.R
@@ -8,7 +8,8 @@ outfile <- basename(infile)
 setup({
   writeLines("Lorem ipsum dolor sit amet, consectetur", infile)
   if (has_pat()) {
-    p1 <<- osf_create_project(title = "osfr-test-files")
+    p1 <<- osf_create_project(title = "osfr-test-files-1")
+    p2 <<- osf_create_project(title = "osfr-test-files-2")
   }
 })
 
@@ -17,6 +18,7 @@ teardown({
 
   if (has_pat()) {
     osf_rm(p1, recursive = TRUE, check = FALSE)
+    osf_rm(p2, recursive = TRUE, check = FALSE)
   }
 })
 
@@ -141,13 +143,15 @@ test_that("a directory can be downloaded as a zip file", {
 })
 
 
-context ("Moving/copying files")
+context("Moving/copying files")
 
 test_that("moving to a destination with an existing file throws an error", {
   expect_error(osf_mv(f1, d1), "Cannot complete action: file or folder")
 })
 
 test_that("moving can overwrite an existing file", {
+  skip_if_no_pat()
+
   f1 <- osf_mv(f1, d1, overwrite = TRUE)
   expect_s3_class(f1, "osf_tbl_file")
 
@@ -158,6 +162,8 @@ test_that("moving can overwrite an existing file", {
 })
 
 test_that("moving destination can be the parent node", {
+  skip_if_no_pat()
+
   f1 <- osf_mv(f1, p1)
   expect_s3_class(f1, "osf_tbl_file")
 
@@ -167,9 +173,9 @@ test_that("moving destination can be the parent node", {
   )
 })
 
-p2 <- osf_create_project("File Tests 2")
-
 test_that("moving destination can be a different node", {
+  skip_if_no_pat()
+
   f1 <- osf_mv(f1, p2)
   expect_match(get_parent_id(f1), as_id(p2))
 })
@@ -186,6 +192,8 @@ test_that("directories can be moved to a sibling directory", {
 })
 
 test_that("moving a parent directory to a child directory errors", {
+  skip_if_no_pat()
+
   parent <- osf_mkdir(p1, "parent")
   child <- osf_mkdir(p1, "parent/child")
   expect_error(
@@ -198,16 +206,22 @@ test_that("moving a parent directory to a child directory errors", {
 context("Deleting files")
 
 test_that("a single file can be deleted", {
+  skip_if_no_pat()
+
   f1 <- osf_refresh(f1)
   expect_true(osf_rm(f1, check = FALSE))
 })
 
 test_that("an empty directory can be deleted", {
+  skip_if_no_pat()
+
   d2 <- osf_mkdir(p1, "d2")
   expect_true(osf_rm(d2, check = FALSE))
 })
 
 test_that("a non-empty directory can be deleted", {
+  skip_if_no_pat()
+
   d3 <- osf_mkdir(p1, "d1/d2/d3")
   expect_true(osf_rm(d3, check = FALSE))
 })
