+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: osfr
 Title: R Interface to OSF
-Version: 0.2.1.9000
+Version: 0.2.2
 Authors@R: c(
     person("Aaron", "Wolen",, "aaron@wolen.com", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0003-2542-2202")),
