### 2019-01-25 #279 @aaronwolen
Hi, @maelle! Thanks for checking out the package and I apologize for the delayed response—I’ve been out traveling most of this week and am still playing catchup.

I made the following changes based on your comments:

* added a code of conduct (centerforopenscience/osfr@e6db814)
* added a note to the README explaining a PAT is necessary to manage projects and upload files, along with a link to the *Authenticating osfr* vignette (centerforopenscience/osfr@68e29f3)
* indicated that contributors need to generate a PAT on the testing server with all scopes enabled (centerforopenscience/osfr@1f601ca, centerforopenscience/osfr@3f520fa)
* reformatted many of the overly long lines (centerforopenscience/osfr@eb9a81b)

    (There are still a few places where I’m breaking the 80 line rule, primarily for errors, warnings, and other messages that aren’t as easily split up.)

Regarding the test failures you encountered, I believe the problem was the tests were run on the main OSF server (`osf.io`) and several of the required resources are located on OSF’s testing server (`test.osf.io`), which is where COS prefers development take place. The [contributing guidelines](https://github.com/CenterForOpenScience/osfr/blob/master/.github/CONTRIBUTING.md) mention you need to set `OSF_SERVER=test` to switch to the testing server, but it’s easy to miss.

This really underscored that the testing infrastructure was a little fragile, so I did some restructuring and added checks to skip certain tests if they’re being run on the production server or if no PAT is defined (centerforopenscience/osfr@81cfe71). It was a big enough change that I bumped the version to 0.2.2.

Thanks again for looking it over and please let me know if you have any other questions or follow-up comments.
