### 2019-01-26 #279 @maelle
Thanks @aaronwolen, great work! Sorry for missing the contributing guidelines! Two small suggestions:

* "Create a .Renviron file in the root of your project directory" or one could edit the existing general `.Renviron`/create one (that's what I did :-) ). You could mention the handy `usethis::edit_r_environ()` function for that.

* Regarding the long lines, you could add `# nolint` at the end of the ones you do not want to split cf https://github.com/jimhester/lintr#project-configuration

I'll now look for reviewers! :smile_cat:
