### 2019-01-26 #279 @aaronwolen
Great suggestions! I did not know about `edit_r_environ()`—that package is just chock full of goodness.
