### 2019-01-28 #279 @maelle
@aaronwolen Great definition of `usethis` :grin:
