### 2019-01-28 #279 @maelle
@HeidiSeibold @cboettig Thanks for accepting to review this package! :smile_cat: Your reviews are due on 2019-02-20. As a reminder here's [our guide for reviewers](https://ropensci.github.io/dev_guide/reviewerguide.html) and [the review template](https://ropensci.github.io/dev_guide/reviewtemplate.html).
