### 2019-01-28 ccbf6c1
Add internal osf_tbl validation function
