### 2019-01-28 fb59b7d
Fix incorrect column name in empty osf_tbls
Closes #88
