+++ README.Rmd
@@ -127,7 +127,7 @@ Big thanks to Rusty Speidel for designing our logo and [Tim Errington][tim] for 
 
 ## Contributing
 
-Check out the [Contributing Guidelines](.github/CONTRIBUTING.md) to get started with osfr development and note that by contributing to this project, you agree to abide by the terms outlined in the [Contributor Code of Conduct](CODE_OF_CONDUCT.md).
+Check out the [Contributing Guidelines](.github/CONTRIBUTING.md) to get started with osfr development and note that by contributing to this project, you agree to abide by the terms outlined in the [Contributor Code of Conduct](.github/CODE_OF_CONDUCT.md).
 
 <!-- links -->
 [osf]: https://osf.io "Open Science Framework"
