### 2019-01-29 6907529
Implement crul's new verb and retry methods
