### 2019-01-29 a6868cd
Bump version and update NEWS
