### 2019-01-29 #279 @aaronwolen
@maelle, a new version of purrr was *just* released that, unfortunately, introduced a bug in osfr. It's fixed in the [dev branch](https://github.com/CenterForOpenScience/osfr/tree/dev) but I wanted verify it's okay to merge the new version now that reviewers have been assigned?
