### 2019-01-29 c600208
Skip two additional tests if no PAT
