### 2019-01-30 #279 @HeidiSeibold
Question: what is the difference between:
`remotes::install_github()`
and
`devtools::install_github()`

The package README is instructing to use the former. Does that matter at all @maelle?
