### 2019-01-30 #279 @maelle
@richfitz said "`devtools::install_github` calls `remotes::install_github` in recent devtools versions (since 2.0.0)" so there shouldn't be any difference. :-)
