### 2019-02-07 #279 @aaronwolen
Hi, @HeidiSeibold! I just noticed your edits and responded below:

> * Documentation on `osf_upload()` says:
> ```
> overwrite    Logical, overwrite an existing file with the same name (default FALSE)? If TRUE, OSF will automatically update the file and record the previous version.
> ```
> What happens for `FALSE`? Does it not upload at all or does it upload as a separate file?

The default behavior is to throw the following error

```error
 Error: File already exists at destination
* Set `overwrite=TRUE` to upload a new version
```

---

> (Btw. the issue for me was that I was using the token from the test server instead of the actual osf, where the project lives)

I've run into this several times myself. I'm considering adding a startup message that informs the user when `OSF_SERVER` is defined and they're running on the test server.

---

> * How do I list all components within a project?

The `n_max` argument determines how many components will be listed (really just limiting the number of calls made to the API). The default is 10 but setting it to `Inf` will return *all* components.

I see now that this behavior is not documented! I'll improve the documentation for this argument in the dev branch. Thanks for pointing that out!

---

Thanks for the questions! I'll keep checking back for more.
