### 2019-02-07 ae56ddf
Inform at startup when OSF_SERVER is defined
(ropensci/software-review/issues/279)
