### 2019-02-07 9646d94
Fix testing server startup message
