### 2019-02-11 #279 @HeidiSeibold
@maelle I have trouble checking

and

because for many examples it does not make sense to run them without modification. I will just check them anyway because I can use and modify the examples so it works for me.
