### 2019-02-11 #279 @HeidiSeibold
Sorry, I'll try to clarify:

Many examples should not be run, because they would upload data to osf and you don't want the same data to be uploaded to osf every time someone tries the package. So there is a dont-run-example, which I modify and then try, so I actually upload useful things to osf.

Same for the vignette.
