### 2019-02-11 #279 @aaronwolen
Thanks for the additional comments, @HeidiSeibold!

> * For one of my projects I do not receive all components, when I ask for it (e.g. `talk_2015_07_user `, compare code to screenshot).

  I wanted to make sure you saw my [previous comment](https://github.com/ropensci/software-review/issues/279#issuecomment-461408478) that mentions using the `n_max` argument to specify the max number of results returned by `osf_ls_nodes()`. I updated the docs to indicate that `n_max = Inf` will retrieve *all* results for a query (CenterForOpenScience/osfr@8ff84f4b).

  I believe this is related to your specific example regarding OSF project <https://osf.io/vk95n/>. The default value for `n_max` is 10, so that's likely why the component *talk_2015_07_user* was not included in the results. Try setting `n_max` to something > 10 or `Inf` and let me know if that resolves your issue. The screenshot shows only 9 components and the project components are private, so I'm unable to verify myself.

---

> * If I understand correctly, your main object in the package is `osf_tbl_node`, which consists of `name`, `id` and `meta`. I would appreciate some more documentation on the `meta` part. What info does it contain?
>
>   * I assumed that `meta` would also contain the info on the components of the project, but this seems to not be the case.
>   * I would like to see a deeper documentation of `osf_tbl`
>   * Maybe also a note on the connection between the weblink and the `id` might be helpful for many users.

We do provide some documentation in [`?osf_tbl`](https://github.com/CenterForOpenScience/osfr/blob/master/R/osf_tbl.R), and mention that `meta` contains *"a list-column that stores the processed response returned by OSF's API"*.

I totally agree it could be more detailed. Please let me know if you have any specific details in mind that you feel are missing, it would really help to highlight some of the areas I'm likely overlooking. For example, your request to note the connection between OSF URLs and `id`s was great—definitely an oversight on our part.

---

> * What is the difference between a project, a component and a directory? I think I have a rough understanding of it, but it might be helpful to write this down somewhere explicitly.

Projects, components, and directories represent the 3 different levels of organization on OSF that can be used to structure a project. OSF actually provides some really great documentation on how these entities compare to each other:

* [Projects](http://help.osf.io/m/projects/l/481539-create-a-project) are the top-level structures on OSF
* [Components](http://help.osf.io/m/projects/l/481998-create-components) are nested within a project and have their own storage buckets and, importantly, privacy settings
* [Folders](http://help.osf.io/m/projectfiles/l/524183-create-and-manage-folders) can be created within projects or components to organize files but they have the same privacy settings as the parent project/component

We reference these links in a few places but they should probably be more front and center in the README, or perhaps in, `?osfr`. That might help users less familiar with OSF get started.

---

> * Will `osfr` in the future also allow interacting with files on e.g. GitHub, which are connected to osf?

Yes! We're definitely planning to add support for the other providers that OSF works with. The plan was to first make sure the package works well/reliably with OSF's native cloud storage, and then include 3rd-part services in a future release.

> (Related to above question, but more of a personal interest question: if I have files connected to osf from GitHub, would the files be lost if GitHub was shut down?)

Adding GitHub to an OSF project just creates a connection between the two services, but the GitHub files remain on GitHub. So if GitHub shuts down, the files would be lost.

You *can* backup external files stored on connected 3rd party service by [registering the project](http://help.osf.io/m/registrations/l/524205-register-your-project).
---

* Are you planing on creating a pkgdown-website?

Planned and deployed: https://centerforopenscience.github.io/osfr
