### 2019-02-11 20c4b3d
Better document the GUID/URL connection and Waterbutler IDs
