### 2019-02-11 e19d8f2
Add simple downloading example to README (#91)
