### 2019-02-13 #279 @HeidiSeibold
Thanks @aaronwolen for your answers. They are very helpful and I hope I am helpful in making your package even better than it is already :wink:

>> For one of my projects I do not receive all components, when I ask for it (e.g. talk_2015_07_user , compare code to screenshot).
>
> I wanted to make sure you saw my previous comment that mentions using the n_max argument to specify the max number of results returned by osf_ls_nodes(). I updated the docs to indicate that n_max = Inf will retrieve all results for a query (CenterForOpenScience/osfr@8ff84f4).

Sorry, yes I did read it but immediately forgot. Works now!

> We do provide some documentation in ?osf_tbl, and mention that meta contains "a list-column that stores the processed response returned by OSF's API".
>
> I totally agree it could be more detailed. Please let me know if you have any specific details in mind that you feel are missing, it would really help to highlight some of the areas I'm likely overlooking. For example, your request to note the connection between OSF URLs and ids was great—definitely an oversight on our part. I tried to make it clearer here: (CenterForOpenScience/osfr@22e8af2)

Yeah I do agree that a bit more detail would be helpful. Some examples:

- Does `meta` always have only one list item `data`?
- What is the meaning of:
    * `attributes`
    * `links`
    * `relationships`
    * potential other items in `meta`/`data`
- Are any infos within these list items helpful for the user? Pick some examples and show the use for them.

> We reference these links in a few places but they should probably be more front and center in the README, or perhaps in, ?osfr. That might help users less familiar with OSF get started.

Yes I agree. These objects are essential for understanding how to interact with OSF. Also an overview on actions which can be done with the objects would be helpful. On OpenML we are using a [cheatsheet](https://github.com/openml/openml-r/blob/master/vignettes/openml-cheatsheet.pdf) with clear visualisations to help users find the functionality they need at each moment. I think OSF has similar features (listing, downloading, etc) so this might be helpful to look at.
