### 2019-02-15 c6ed366
Verify match is exact when recursing paths (fixes #95)
