### 2019-02-15 6f560f1
Bump to dev version and update news
