### 2019-02-15 586a959
Note that modification time determines osf_ls_*() order (closes #94)
