### 2019-02-18 #279 @aaronwolen
Thanks again, @HeidiSeibold. I'll be sure to include more details about the `meta` column in the next release. We're also planning to add a function that allows users to easily retrieve useful information about OSF objects from `meta` but I'm still working on the implementation.

I also think a cheatsheet is a great idea, it's definitely on the todo list. The OpenML cheatsheet is a really nice example, thanks for sharing!
