### 2019-02-18 7f2753d
Fix verbose argument for osf_upload
