### 2019-02-22 9161578
tweak instructions to be more explicit
