### 2019-02-22 #279 @cboettig
sorry I'm late!  😊

---

Really professionally developed package here.  All the tests and examples ran smoothly for me. The documentation, the unit tests, and the code were easy to ready and understand and look very thoughtfully implemented.  I have a few minor comments, some conceptual questions about the design, and some recommendations on regarding the documentation.

I think this is a very valuable package and am looking forward to testing it out in real use cases in my lab group over the coming weeks, since I think it could serve many of our needs that aren't met by other existing utilities.  exciting stuff!

#### Conceptual questions:

- Are there any operations from the API that can be supported without authentication?  In particular, I think the most useful feature would be the ability to download public files without an API key.  These can be downloaded in the browser without logging in, so this would seem comparable.  (Unfortunately from the browser end, I don's seem to be able to get a hard URL for the download since the browser interface itself doesn't seem to be RESTful...)

- While I value the low-level API, manually recreating directory structure in the `create_project` example becomes tedious for a large(r) project.  Rather than indicate the names of each file and folder, it would be very nice to have a higher level interface that could simply  import the files and folders of an existing project; potentially indicating which files to exclude (a la `.gitignore`).  This would be straight forward enough for a user to implement with the existing package, but would make more sense at this level.

- Similarly, I imagine some users might want to import a project from GitHub. I saw http://help.osf.io/m/addons/l/837075-connect-github-to-a-project, but would be nice to see a scripted way to do this from `osfr` if possible; at least after a one-time online enable.

- While this may be beyond what the API can provide, I think it would also be really nice to be able to control user permissions from the command line.  In particular, it would be nice to be able to grant other users permission to view or view & edit files automatically from the R package by specifying user IDs.  (For instance, this can be done in DataONE, though the downside to their system imho is the shortlived tokens make daily use for private / in development work inpractical.)

- Are there plans to extend package functionality beyond the current scope of just file/folder/node management?  The web interface even without extensions is so much richer that this currently feels somewhat narrow, (though is certainly the most useful bit.)  For instance, it might be nice to get download statistics etc for a project or file.

- I would appreciate more guidance up-front on file size limits, project size limits, and permission management in the Getting Started Vignette, etc.

### Software design

- The one issue I see is the lack of support to pass a token to the function call.  Typical design patterns I like here provide the option of specifying a token as an optional argument (e.g. `.token` in `gh` package, also see `usethis` & `devtools` examples), giving the user an alternative to setting the environmental variable.  This can be useful for more complex authentication designs (scripts authenticating in multiple users) or for developers, or just more convienient for a one-off function call.

### nits / misc

- README.Rmd/md:

  - Why `Sys.unsetenv(c("OSF_SERVER", "OSF_PAT"))`?
  - `r_project <- osf_retrieve_node("e81xl")` 404 error: invalid token (with & without my token).  Presumably not meant to be run with test account tokens?

  - Consider linking to the (very nice) vignettes directly.

  - very nice `CONTRIBUTING.md`, minor tweaks to wording in a PR!
