### 2019-02-22 #279 @csoderberg
Thanks for the comments @cboettig.

> Are there any operations from the API that can be supported without authentication?  In particular, I think the most useful feature would be the ability to download public files without an API key.  These can be downloaded in the browser without logging in, so this would seem comparable.  (Unfortunately from the browser end, I don's seem to be able to get a hard URL for the download since the browser interface itself doesn't seem to be RESTful...)

You should only need authentication to make edits to projects, or to download private files. Any public node or file can be accessed without authentication. So `osf_retrieve_file('fgjvw') %>% osf_download` should work without a key. Are you getting an error, or do we just need to make it more clear in the documentation what can be done without authentication?

> Similarly, I imagine some users might want to import a project from GitHub. I saw http://help.osf.io/m/addons/l/837075-connect-github-to-a-project, but would be nice to see a scripted way to do this from `osfr` if possible; at least after a one-time online enable.

You is technically possible through the API, but those endpoints aren't particularly stable right now. We also decided to restrict the initial release of the package to just work just with OSF storage, and potentially tackle add-ons in future releases if there is community requests for this.

> * While this may be beyond what the API can provide, I think it would also be really nice to be able to control user permissions from the command line.  In particular, it would be nice to be able to grant other users permission to view or view & edit files automatically from the R package by specifying user IDs.  (For instance, this can be done in DataONE, though the downside to their system imho is the shortlived tokens make daily use for private / in development work inpractical.)
* Are there plans to extend package functionality beyond the current scope of just file/folder/node management?  The web interface even without extensions is so much richer that this currently feels somewhat narrow, (though is certainly the most useful bit.)  For instance, it might be nice to get download statistics etc for a project or file.

These two point are related. It's possible to do the first point through the API. The [API](https://developer.osf.io) is really extensive, and as you correctly pointed out enables much richer functionality than the package currently covers. I think adding contributor management is a great suggestion for a future version of the package, and we do have plans to extend out the functionality in later releases. Given the long development history fo this package, and our desire to have a stable version of the package released and on CRAN, we decided to focus in on a smaller set of core features to start in order to limit the initial scope to something manageable for the first official version of the the package. For future versions, I know we have a few features in the back of our minds that we'd like to include, and are also very open to community feedback about what features would be most useful to them to help prioritize when goes out in future releases.

> * `r_project <- osf_retrieve_node("e81xl")` 404 error: invalid token (with & without my token).  Presumably not meant to be run with test account tokens?

I'm not able to recreate this error. Were you by chance pointing towards the test.osf.io API endpoint, rather than the main osf.io API when you made that call? The 'usage example' and the 'vignette' use example public projects form OSF, rather than our test server.
