### 2019-02-22 #279 @cboettig
Re 404 error: thanks, yes, I'm able to get this to work in a fresh session where I have not set `OSF_SERVER=test` in .Renviron.  I am not able to get it to work though in the session where I have set `test` in `.Renviron`, even after I run `Sys.unsetenv(c("OSF_SERVER", "OSF_PAT"))` though, which I would have thought would be sufficient.  Can you see if you get a 404 if you set `test` in the `.Renviron` but then unload it?

Aside from this weird technical glitch though, this is great, very nice to be able to download public files without a token!

- Regarding future development plans covering more of the API: that all sounds great too.  It might be nice if these were included as Issues and/or Milestones in the package repo?

I still think it would be nice to support creation of a project that has multiple files and folders as a one-liner function call, rather than the very manual current example.  This task is the feature-example shown in the readme and getting started docs, but takes 5 function calls to upload one data file in one project.  I'd love something more analogous to `git init`, `git add -A`/`git commit`, `git push` style to upload a larger project with many files.  This could be built out entirely on the R end with just a bit of `dir.list()` or `fs::dir_ls()` style commands and would be helpful in getting larger projects started.  Automating the process for many files is probably one of the more compelling reasons to create a package manually in the first place instead of using the web interface, and I'd love to review such a capability here.  Nevertheless, I appreciate the goal of not holding up your submission to CRAN so I'd be happy to see these things just on the roadmap / milestones; clearly the core functionality is both solid and valuable as is, so this is just a suggestion which I think would improve the first-impressions of usability for real-life cases.
