### 2019-02-26 #279 @aaronwolen
First of all, thanks so much for the thoughtful review and kind words, @cboettig. It’s great to hear that osfr might be useful in your lab.

Courtney did a nice job addressing your questions concerning authentication and supporting add-ons, so I’ll start with your suggestion about adding a higher level interface for adding an existing project to OSF:

> While I value the low-level API, manually recreating directory structure in the `create_project` example becomes tedious for a large(r) project.  Rather than indicate the names of each file and folder, it would be very nice to have a higher level interface that could simply  import the files and folders of an existing project; potentially indicating which files to exclude (a la `.gitignore`).  This would be straight forward enough for a user to implement with the existing package, but would make more sense at this level.

I totally agree the number of steps required for certain tasks is tedious at the moment. Now that we have a core set of atomic functions for performing (what I assume are) the most common OSF operations, I plan to focus on making typical workflows more convenient. Your example of uploading/importing an existing project to OSF is a perfect example.

I’ve been working on updated versions of the upload and download functions that support multiple files/directories. This *should* make the process of adding an existing project to OSF slightly less painful since you could upload your entire project directory at once:

```r
my_proj <- osf_create_project("My Existing Project")
osf_upload(my_proj, path = getwd())
```

Love your suggestion for a `.gitignore`-style feature! It would be a great complement to the above workflow, making it possible to implement specific filters without requiring any additional arguments that might complicate the interface.

Do you think the combination of these features would do the trick?

---

> Are there plans to extend package functionality beyond the current scope of just file/folder/node management?  The web interface even without extensions is so much richer that this currently feels somewhat narrow, (though is certainly the most useful bit.)  For instance, it might be nice to get download statistics etc for a project or file.

In terms of providing access to the metadata that OSF provides about projects and files, that’s absolutely on the roadmap. Much of that info is already present within the `meta` column, so a short-term fix will be to better document what’s hiding in there (as suggested by @HeidiSeibold), more advanced users can pluck out what they need. After that, I’m planning to include a new function for adding relevant attributes to an `osf_tbl`. Here’s an example:

```r
pysch_rp <- osf_retrieve_node("ezum7")

psych_rp %>%
  osf_get(c("category", "date_created", "tags"))

# # A tibble: 1 x 6
#   name                                                    id    category      date_created        tags       meta
#   <chr>                                                   <chr> <chr>         <dttm>              <list>     <list>
# 1 Estimating the Reproducibility of Psychological Science ezum7 communication 2015-07-20 17:42:16 <list [3]> <list [3]>
```

This will also coincide with a new `ls`-like argument for `osf_ls_files()` and `osf_ls_nodes()` to include more information in their output, e.g.

```r
osf_ls_files(psych_rp, info = "extra")

# # A tibble: 10 x 7
#    name                              id                       kind   size      modified            downloads meta
#    <chr>                             <chr>                    <chr>  <chr>     <dttm>                  <int> <list>
#  1 RPP_Supplement.pdf                55f1a8d18c5e4a093bc70500 file   822.89 kB 2015-09-18 01:11:16       517 <list [3]>
#  2 RPP_SI_Figures                    55ad3c7e8c5e4a527e115b8d folder 0 B       2015-09-18 01:11:16         0 <list [3]>
#  3 RPP_Appendices.pdf                55ad337b8c5e4a5277114f47 file   188.72 kB 2015-09-18 01:11:16       285 <list [3]>
#  4 RPP_Figure3.pdf                   55ad3c0b8c5e4a527c115553 file   27.83 kB  2015-09-18 01:11:16       130 <list [3]>
#  5 RPP_Figure1.pdf                   55ad3c0b8c5e4a527e115b7b file   16.59 kB  2015-09-18 01:11:16       191 <list [3]>
#  6 RPP_Figure2.png                   55ad3c0c8c5e4a527e115b7e file   51.27 kB  2015-09-18 01:11:16       124 <list [3]>
#  7 RPP_Table_1.pdf                   55ad3c168c5e4a527e115b81 file   31.23 kB  2015-09-18 01:11:16       168 <list [3]>
#  8 RPP_Table_2.pdf                   55ad3c178c5e4a527711503c file   43.79 kB  2015-09-18 01:11:16       167 <list [3]>
#  9 RPP_Figure3_compressed.jpg        55e05f118c5e4a2f9b1a72d7 file   75.96 kB  2015-09-18 01:11:16        93 <list [3]>
# 10 RPP_Figure1_panelA_compressed.jpg 55e0b3408c5e4a6d5cadaede file   19.86 kB  2015-09-18 01:11:16        87 <list [3]>
```

---

> The one issue I see is the lack of support to pass a token to the function call.  Typical design patterns I like here provide the option of specifying a token as an optional argument (e.g. `.token` in `gh` package, also see `usethis` & `devtools` examples), giving the user an alternative to setting the environmental variable.  This can be useful for more complex authentication designs (scripts authenticating in multiple users) or for developers, or just more convenient for a one-off function call.

It was something I considered initially but ultimately decided against, figuring most users would opt to authenticate using the env variable, and those who don't could still pass their PAT directly to `osf_auth()`. Do you think that would suffice in the multi-user scenarios you're envisioning?

---

> * Why `Sys.unsetenv(c("OSF_SERVER", "OSF_PAT"))`?

Code chunks in the README need to access public resources on OSF. Since the README is usually only rendered in a dev environment where `OSF_SERVER=test` and the `OSF_PAT` was generated on <test.osf.io>, this temporarily unsets those variables to avoid 404 errors. I don't *think* this has any side effects...

As to the 404 you're experiencing, I'm not able to recreate it on macOS running R 3.5.1. Could you give me a little more info your setup?

---

BTW, happy to keep this dialogue going post-review if you have additional ideas about how to make the package more useful to your lab.
