### 2019-02-26 #279 @cboettig
@aaronwolen This is great.

Regarding uploading a project's files & directory structure in one go with

```r
my_proj <- osf_create_project("My Existing Project")
osf_upload(my_proj, path = getwd())
```

that sounds like a most welcome addition!  cc'ing @karinorman who has been battle-testing this with some complex project structures, we'll try and give that a test-drive.  The previously manual approach was a bit of a show-stopper for us at first pass.

Re 404, I see that also on MacOS (high Sierra) on R 3.5.2; but it seems to be a weird side-effect of using  `Sys.unsetenv(c("OSF_SERVER", "OSF_PAT"))` interactively?  Anyway, seems related to the test setup and unlikely to be an issue for users, so probably not worth diving into.

Regarding the token, I'm not quite clear what the workflow would look like if I wanted to say, upload to three different projects with distinct tokens in the same script; I guess just call `osf_auth()` each time before each function call?  Relying on side-effects instead of being able to pass all parameters if need be undercuts a functional-programming workflow, though in practice it's probably not a big deal.  I certainly agree that most users will just want to set `.Renviron` and that the functions should check for the tokens there if no `.token` argument is given, I just like functions to also be able to run as pure functions where all arguments are specified as arguments.  Happy to let that go here.

@maelle Aaron, Courtney and co have satisfied all my issues here, so I'm happy to approve this package :+1: and continue any further discussion with them about full-directory uploads etc in the issues tracker for the `osfr` project itself instead.  Very nice package here and a great contribution to the R community.
