+++ R/api-endpoints-wb.R
@@ -55,13 +55,18 @@
 #'   folders.
 #'
 #' @noRd
-.wb_download <- function(id, fid, path, type, zip = FALSE) {
+.wb_download <- function(id, fid, path, type, zip = FALSE, verbose = FALSE) {
   type <- match.arg(type, c("file", "folder"))
   query <- list()
   if (zip) query$zip <- ""
   api_path <- .wb_api_path(id, fid, type = type)
   res <- .wb_request("get", api_path, query, disk = path)
-  if (res$status_code == 200) return(TRUE)
+
+  if (res$status_code == 200) {
+    if (verbose) message(sprintf("Downloaded OSF %s to %s", type, path))
+    return(TRUE)
+  }
+
   if (res$status_code == 404) {
     msg <- sprintf("The requested %s (%s) could not be found in node `%s`",
                    type, fid, id)
