### 2019-02-27 3b6fe57
Add function for unzipping multiple files
