+++ NAMESPACE
@@ -44,10 +44,12 @@ importFrom(crul,url_parse)
 importFrom(fs,is_dir)
 importFrom(fs,is_file)
 importFrom(fs,path_dir)
+importFrom(fs,path_ext_remove)
 importFrom(fs,path_ext_set)
 importFrom(jsonlite,fromJSON)
 importFrom(purrr,"%>%")
 importFrom(purrr,"%||%")
+importFrom(purrr,discard)
 importFrom(purrr,imap_chr)
 importFrom(purrr,map)
 importFrom(purrr,map_chr)
