### 2019-02-27 737a0a5
Add tests for unzip_files()
