### 2019-02-27 622423b
Add decompress option to osf_download()
