### 2019-02-27 479d1de
Account for embedded directories when unzipping
