### 2019-02-27 83b5960
Add missing decompress argument
