### 2019-02-27 7d16045
Fix local_paths for downloaded files in pwd
