### 2019-02-27 5be960c
Fix unzip_files() ability to find existing files
