+++ R/utils-unzip.R
@@ -9,26 +9,35 @@
 #'
 #' @return Invisibly returns a named list for each `zipfile` containing the
 #'   unzipped file paths
-#' @importFrom purrr discard
 #' @noRd
 unzip_files <- function(zipfiles, overwrite = FALSE) {
   stopifnot(all(fs::path_ext(zipfiles) == "zip"))
   stopifnot(all(file.exists(zipfiles)))
 
   zipped_files <- map(zipfiles, ~ unzip(.x, list = TRUE)$Name)
+  extract_dirs <- map(zipfiles, fs::path_ext_remove)
+
+  # unzip to pwd if the exdir is embedded in the zip file
+  redundant <- purrr::map2_lgl(
+    .x = zipped_files,
+    .y = extract_dirs,
+    ~ fs::path_has_parent(.x, .y)
+  )
+  extract_dirs <- ifelse(redundant, ".", extract_dirs)
 
   # remove existing files to prevent overwriting
   if (isFALSE(overwrite)) {
-    zipped_files <- map(zipped_files, purrr::discard, .p = file.exists)
+    zipped_files <- Map(
+      function(file, dir) {
+        path <- file.path(dir, file)
+        file[!file.exists(path)]
+      },
+      file = zipped_files,
+      dir = extract_dirs
+    )
   }
 
-  # unzip to a directory named for the zipfile
-  exdirs <- map(zipfiles, fs::path_ext_remove)
-  # unzip to pwd if the exdir is embedded in the zip file
-  redundant <- purrr::map2_lgl(zipped_files, exdirs, ~ fs::path_has_parent(.x, .y))
-  exdirs <- ifelse(redundant, ".", exdirs)
-
-  unzipped <- Map(unzip, zipfile = zipfiles, files = zipped_files, exdir = exdirs)
+  unzipped <- Map(unzip, zipfile = zipfiles, files = zipped_files, exdir = extract_dirs)
 
   unlink(zipfiles)
   invisible(unzipped)
