+++ NAMESPACE
@@ -49,7 +49,6 @@ importFrom(fs,path_ext_set)
 importFrom(jsonlite,fromJSON)
 importFrom(purrr,"%>%")
 importFrom(purrr,"%||%")
-importFrom(purrr,discard)
 importFrom(purrr,imap_chr)
 importFrom(purrr,map)
 importFrom(purrr,map_chr)
