### 2019-02-27 e1c5563
Update download tests
