+++ NAMESPACE
@@ -66,3 +66,4 @@ importFrom(utils,head)
 importFrom(utils,menu)
 importFrom(utils,modifyList)
 importFrom(utils,tail)
+importFrom(utils,unzip)
