### 2019-03-01 #279 @cboettig
Just a quick correction to my comments above: looks like I jumped the gun here and the "upload local directory+files) feature is still under development, correct?

 It would be nice to give it a test run and see it show up in the documentation (perhaps in place of the more tedious current example).    Happy to test-drive it from a dev branch if you point us in the right direction
