### 2019-03-01 #279 @cboettig
Thanks @aaronwolen this sounds great.  I think we'd get a lot of use out of multi file uploading and downloading too -- looking forward to it.
