### 2019-03-09 9cff6f3
Download message printed before initiating download
