+++ tests/testthat/test-files.R
@@ -115,72 +115,6 @@ test_that("messages are printed with `verbose` enabled", {
 })
 
 
-context("Downloading")
-test_that("a file can be downloaded from a project", {
-  skip_if_no_pat()
-
-  out <- osf_download(f1)
-  expect_s3_class(out, "osf_tbl_file")
-  expect_identical(out$local_path, outfile)
-  expect_true(file.exists(outfile))
-})
-
-test_that("an existing file won't be overwritten", {
-  skip_if_no_pat()
-
-  expect_error(
-    osf_download(f1),
-    "A file exists at the specified"
-  )
-  expect_s3_class(
-    osf_download(f1, overwrite = TRUE),
-    "osf_tbl_file"
-  )
-})
-
-test_that("a non-existant path throws an error", {
-  skip_if_no_pat()
-
-  expect_error(
-    osf_download(f1, path = "ddd"),
-    "`path` must point to an existing local directory.")
-})
-
-test_that("a file can be downloaded from a directory", {
-  skip_if_no_pat()
-
-  unlink(outfile)
-  out <- osf_download(f2)
-  expect_s3_class(out, "osf_tbl_file")
-  expect_identical(out$local_path, outfile)
-  expect_true(file.exists(outfile))
-})
-
-test_that("a directory can be downloaded as a zip file", {
-  skip_if_no_pat()
-
-  zipfile <- file.path(tempdir(), paste0(d1$name, ".zip"))
-  out <- osf_download(d1, path = tempdir(), decompress = FALSE)
-  expect_s3_class(out, "osf_tbl_file")
-  expect_true(file.exists(zipfile))
-  unlink(zipfile)
-})
-
-test_that("a downloaded directory is unzipped", {
-  skip_if_no_pat()
-
-  d1_files <- osf_ls_files(d1, n_max = Inf)
-  outdir <- file.path(tempdir(), d1$name)
-
-  out <- osf_download(d1, path = tempdir())
-  expect_s3_class(out, "osf_tbl_file")
-
-  expect_true(dir.exists(outdir))
-  expect_true(all(file.exists(file.path(outdir, d1_files$name))))
-})
-
-
-
 context("Moving/copying files")
 
 test_that("moving to a destination with an existing file throws an error", {
