### 2019-03-09 3e46884
Add rudimentary request logging
