### 2019-03-09 0339dc2
Redirect download test to tempdir
