### 2019-03-09 5139ffe
Unzipping ignore embedded subdirectories
The was originally added to accomodate tests that were embedding
subdirectory paths within the zip file. Testing indicates this should
never happen with zip files created by OSF
