### 2019-03-09 b489bdf
Refactor unzip_files() tests
