### 2019-03-10 b6501b2
Log 5xx errors and retry attempts
