+++ R/utils-logging.R
@@ -1,8 +1,6 @@
 log_response <- function(res) {
-  logfile <- getOption("osfr.log")
-  if (is.null(logfile)) return(res)
-  logger::log_appender(logger::appender_file(logfile))
-  logger::log_formatter(logger::formatter_sprintf)
-  logger::log_info(fmt = "%s %s", toupper(res$method), res$request$url$url)
+  if (!is.null(getOption("osfr.log"))) {
+    logger::log_info(fmt = "%s %s", toupper(res$method), res$request$url$url)
+  }
   res
 }
