+++ R/api-client-wb.R
@@ -67,5 +67,10 @@
   method <- match.arg(method, c("get", "put", "patch", "post", "delete"))
   cli <- .wb_cli()
   res <- cli$verb(method, path, query, body = body, ...)
-  log_response(res)
+
+  if (!is.null(getOption("osfr.log"))) {
+    logger::log_info(fmt = "%s %s", toupper(res$method), res$request$url$url)
+  }
+
+  res
 }
