### 2019-03-10 a48b415
Remove name attached only to singleton OSF entities
