### 2019-03-29 0a21631
Add details about osf_tbl meta column
