+++ R/osf_tbl.R
@@ -4,17 +4,14 @@
 #' data frames based on the [tibble][tibble::tibble-package] class. See below
 #' for additional details.
 #'
-#' Each row of an `osf_tbl` represents a single OSF entity, which could be a
+#' Each row of an `osf_tbl` represents a single OSF entity. This could be a
 #' user, project, component, directory, or file. An `osf_tbl` must include
-#' the following 3 variables:
+#' the following 3 columns:
 #'
-#' 1. `name`: the name or title of the entity.
-#' 2. `id`: the unique identifier assigned to the entity.
+#' 1. `name`: indicates the name or title of the entity.
+#' 2. `id`: the unique identifier assigned by OSF.
 #' 3. `meta`: a list-column that stores the processed response returned by OSF's
-#' API.
-#'
-#' The `meta` column is primarily intended for use by `osfr`'s functions and
-#' should not be altered by users.
+#'    API. See the *Meta column* section below for more information.
 #'
 #' @section Subclasses:
 #'
@@ -32,6 +29,22 @@
 #' component must have a parent node (i.e., must be a sub-component of another
 #' project or component). Because projects and components are functionally
 #' identical, osfr uses the same [`osf_tbl_node`] class to represent both.
+#'
+#' @section Meta column:
+#'
+#' The `meta` column contains all of the information returned from OSF's API for
+#' a single entity, structed as a named list with 3 elements:
+#'
+#' 1. `attributes` contains metadata about the entity (e.g., names,
+#'    descriptions, tags, etc).
+#' 2. `links` contains urls to API endpoints with alternative representations of
+#'    the entity or actions that may be performed on the entity.
+#' 3. `relationships` contains URLs to other entities with relationships to the
+#'    entity (e.g., collaborators attached to a project).
+#'
+#' This information is critical for `osfr`'s internal functions and should not
+#' be altered by users. For even more information about these elements, see
+#' [OSF's API documentation](https://developer.osf.io/#tag/Entities-and-Entity-Collections).
 #'
 #' @section Acknowledgments:
 #'
