### 2019-03-29 76d693b
Update news and bump version
