### 2019-03-29 #279 @aaronwolen
Thanks for following-up, @maelle! Apologies for the delay. Since initiating this review I've changed employers and moved to a new state... so it's been a hectic month 😃.

I just pushed v0.2.4 which should include all of the suggested changes with the exception of  multi-file uploading and downloading. That's definitely next on the list but will require a little more time to get it polished ready for release.
