### 2019-05-16 8d036e7
Use crul hooks to simplify logging
