+++ R/utils-logging.R
@@ -1,6 +1,25 @@
 log_response <- function(res) {
   if (!is.null(getOption("osfr.log"))) {
-    logger::log_info(fmt = "%s %s", toupper(res$method), res$request$url$url)
+    times <- res$times[res$times > 0]
+
+    msg <- sprintf(fmt = "Status %s - %s - [%s]",
+      res$status_code,
+      res$url,
+      paste0(sprintf("%s:%f", names(times), times), collapse = ", ")
+    )
+
+    if (res$status_code >= 500) {
+      msg <- paste(msg, rawToChar(res$content), sep = " - ")
+      logger::log_error("%s", msg)
+    } else {
+      logger::log_info("%s", msg)
+    }
   }
-  res
+}
+
+
+log_request <- function(req) {
+  if (!is.null(getOption("osfr.log"))) {
+    logger::log_info(fmt = "%s %s", toupper(req$method), req$url$url)
+  }
 }
