### 2019-05-17 #279 @aaronwolen
Hey @maelle, Things are coming along, albeit a little more slowly than expected (my side-project time has been limited recently). So far I've got multi-file downloads working well 👇

![Screen capture demonstrating multi-file downloading with osfr.](https://user-images.githubusercontent.com/1067915/57950625-6de65e00-78ad-11e9-8525-2045b95acfee.gif)

Uploading is still a WIP but I hope to have it completed in the near future 🤞.
