### 2019-05-17 eddf566
Improve osf_upload error message (#102)
