+++ R/utils-checks.R
@@ -18,3 +18,20 @@ check_category <- function(x) {
   )
   match.arg(tolower(x), categories)
 }
+
+
+#' Verify local file paths exist
+#' @param files Character vector containing paths for one or more local files
+#' @noRd
+check_files <- function(files) {
+  found <- fs::file_exists(files)
+  if (any(!found)) {
+    abort(paste0(
+      "Can't find find following files/folders included in `path`:\n",
+      paste0(
+        sprintf("  * %s", names(Filter(isFALSE, found))),
+        collapse = "\n")
+    ))
+  }
+  files
+}
