+++ R/osf_upload.R
@@ -1,17 +1,41 @@
-#' Upload files to OSF
+#' Upload files and directories to OSF
+#'
+#' Local files and directories can be uploaded to a project, component or directory on OSF.
+#'
+#' @section A note about synchronization:
+#' While `osf_download()` and `osf_upload()` can be used to conveniently shuttle
+#' files back and forth between OSF and your local machine, it's important to
+#' note that **they are not file synchronization functions**. In contrast to
+#' something like [`rsync`](https://rsync.samba.org),
+#' `osf_download()`/`osf_upload()` do not take into account file contents or
+#' file modification times. Whether you're uploading or downloading, if
+#' `overwrite=TRUE`, osfr will overwrite an existing file regradless of whether
+#' the existing file is the more recent copy. You have been warned.
+#'
+#' @section Uploading to subdirectories:
+#' If you want to upload to an existing directory on OSF, you will first need to retrieve it. For example, if the project `proj` has a subdirectory, `rawdata/`, nested within a top-level directory, `data/`, `osf_ls_files()` could be used to retrieve `rawdata/` directly
+#'
+#' ```
+#' rawdata_dir <- osf_ls_files(proj, path = "data", pattern = "rawdata")
+#' osf_upload(rawdata_dir, path = "my-file.txt")
+#' ```
 #'
 #' @param x an [`osf_tbl_node`] with a single project or
-#'   component, or an [`osf_tbl_file`] with a single directory
-#' @param path Path to a local file. Ensure the file has a proper
-#'   file extension (e.g., `.docx`) to ensure it's rendered properly on OSF.
-#' @param name Name of the uploaded file (if `NULL`, `basename(path)`
-#'   will be used).
+#'   component, or an [`osf_tbl_file`] with a single directory. This is the OSF
+#'   bucket to which your files will be uploaded.
+#' @param path A character vector of filepaths to upload to `x` on OSF. Only
+#'   paths pointing to existing files and/or directories are uploaded. Any other
+#'   file types (e.g., symlinks) are silently ignored.
 #' @param overwrite Logical, overwrite an existing file with the same name
 #'   (default `FALSE`)? If `TRUE`, OSF will automatically update the file and
-#'   record the previous version.
+#'   record the previous version. If `FALSE`, a warning will be issued that the
+#'   local file was *not* uploaded and the *existing* version of the file on OSF
+#'   is returned.
 #' @template verbose
 #'
-#' @return an [`osf_tbl_file`] containing uploaded file
+#' @return an [`osf_tbl_file`] containing only the files/directories uploaded to
+#'   the root-level of `x`. It will not include the files/directories that were
+#'   uploaded to subdirectories within `x`.
 #'
 #' @examples
 #' \dontrun{
@@ -40,14 +64,8 @@
 osf_upload <-
   function(x,
            path,
-           name = NULL,
            overwrite = FALSE,
            verbose = FALSE) {
-
-  if (!file.exists(path)) abort(sprintf("Can't find file:\n %s", path))
-  if (is_dir(path)) {
-    abort("`path` must point to a file\n* Uploading directories is not supported")
-  }
   UseMethod("osf_upload")
 }
 
@@ -55,37 +73,22 @@ osf_upload <-
 osf_upload.osf_tbl_node <-
   function(x,
            path,
-           name = NULL,
            overwrite = FALSE,
            verbose = FALSE) {
 
-  if (is.null(name)) name <- basename(path)
-  name <- check_upload_name(name)
+  path <- check_files(path)
[…]
