### 2019-05-23 2880d82
Update news and version
