### 2019-05-25 49c7eca
Import memoise
