### 2019-05-25 8bf816d
Record guids required by tests in a dcf file
This avoids needing to manually update the test GUIDs if the OSF project is recreated
