### 2019-05-25 0eb430a
Tweak log format
Make status uppercase to match curl requests
