### 2019-05-25 c11f54f
Minor cleanup of recursive upload
