### 2019-05-25 295825a
Bump dev version and update docs
