+++ NAMESPACE
@@ -41,12 +41,19 @@ export(osf_upload)
 importFrom(crul,HttpClient)
 importFrom(crul,upload)
 importFrom(crul,url_parse)
+importFrom(fs,dir_exists)
+importFrom(fs,dir_walk)
+importFrom(fs,file_exists)
+importFrom(fs,file_info)
 importFrom(fs,is_dir)
-importFrom(fs,is_file)
-importFrom(fs,path_dir)
 importFrom(fs,path_ext_remove)
 importFrom(fs,path_ext_set)
+importFrom(fs,path_has_parent)
+importFrom(fs,path_rel)
+importFrom(fs,path_split)
 importFrom(jsonlite,fromJSON)
+importFrom(memoise,forget)
+importFrom(memoise,memoise)
 importFrom(purrr,"%>%")
 importFrom(purrr,"%||%")
 importFrom(purrr,imap_chr)
