### 2019-05-25 464fd6d
Cleaner method for loading required guids
