+++ tests/testthat/test-osf_ls.R
@@ -6,9 +6,10 @@ setup({
   # Retrieve public OSF project and components required for tests
   # (created using data-raw/create-test-project.R)
   if (on_test_server()) {
-    p1 <<- osf_retrieve_node("brfza")
-    c1 <<- osf_retrieve_node("rxwhk")
-    d1 <<- osf_retrieve_file("5c34b68a44cd030016942349")
+    guids <- get_guids()
+    p1 <<- osf_retrieve_node(guids[, "p1"])
+    c1 <<- osf_retrieve_node(guids[, "c1"])
+    d1 <<- osf_retrieve_file(guids[, "d1"])
   }
 })
 
