+++ tests/testthat/test-downloading.R
@@ -6,18 +6,7 @@ setup({
   # Retrieve public OSF project and components required for tests
   # (created using data-raw/create-test-project.R)
   if (on_test_server()) {
-
-    guid_file <- list.files(
-      path = rprojroot::find_testthat_root_file(),
-      pattern = "test-guids.dcf",
-      recursive = TRUE,
-      full.names = TRUE
-    )
-
-    guids <- read.dcf(guid_file)
-
-    dir(rprojroot::find_testthat_root_file(), "test-guids.dcf", recursive = TRUE)
-
+    guids <- get_guids()
     p1 <<- osf_retrieve_node(guids[, "p1"])
     c1 <<- osf_retrieve_node(guids[, "c1"])
     f1 <<- osf_retrieve_file(guids[, "f1"])
