### 2019-05-25 c055269
Fix bug in guid test helper
