### 2019-05-26 809fa9c
Overhaul osf_upload's docs
