### 2019-05-26 b2303a2
Further improvement to osf_upload docs
