### 2019-05-26 8d77843
Use on.exit to forget memoised function
