### 2019-05-26 9d1e3e8
Clarify osf_upload's returns
