### 2019-05-26 8d30f5d
Basic tests for multifile uploads
