### 2019-05-27 d4bc7af
recurse_path ignores leading slashes or dots in path
Prevents searching OSF for folderns named "/" or "."
