### 2019-05-27 f9ca68b
Remove unused test variable
