### 2019-05-27 b49e092
Remove extra context
