+++ DESCRIPTION
@@ -17,7 +17,7 @@ Imports:
     purrr,
     tibble,
     rlang,
-    fs,
+    fs (>= 1.3.0),
     memoise
 License: Apache License 2.0 | file LICENSE
 URL: 
