### 2019-05-27 e012fdc
Add retry method to waterbutler client
