### 2019-06-01 53557d8
Consolidate client building functions
