+++ R/zzz.R
@@ -1,6 +1,3 @@
-.osf_api_version <- 2.8
-.wb_api_version  <- 1
-
 .onLoad <- function(libname, pkgname) {
 
   # record personal access token
