### 2019-06-01 475c153
API versions are now hardcoded in the client constructors
`version` is now a required argument for .build_client()

No longer need:
* global .(osf|wb)_api_version variables
* .osf_api_path()
