+++ R/osf_ls_files.R
@@ -97,7 +97,7 @@ osf_ls_files.osf_tbl_file <-
   # manually construct the API path because the 'files' endpoint for nodes lists
   # the enabled storage providers
   api_path <- switch(class(x)[1],
-    osf_tbl_node = .osf_api_path(sprintf("nodes/%s/files/osfstorage/", as_id(x))),
+    osf_tbl_node = sprintf("nodes/%s/files/osfstorage/", as_id(x)),
     osf_tbl_file = crul::url_parse(get_relation(x, "files"))$path
   )
 
