### 2019-06-01 7f1d48b
Fix new bug creating subdirectories
