+++ R/api-client-osf.R
@@ -22,12 +22,11 @@ user_agent <- function(agent = "osfr") {
            ...) {
 
   method <- match.arg(method, c("get", "put", "patch", "post", "delete"))
-  path <- url_path(paste0("v", floor(version)), path)
   cli <- .build_client(api = "osf", encode = "json", version = version)
 
   cli$retry(
     method,
-    path,
+    prepend_version(path, version),
     query,
     body = body,
     times = 3,
