### 2019-06-01 8127d51
Separate uploading and moving tests
Previously moving tests would operate on the OSF files generated by
the create-test-proejct.R script. If these tests failed, a file
required for other tests would be gone, causing additional errors.
