### 2019-06-01 7857368
First pass at file transfer progress bars
