+++ DESCRIPTION
@@ -18,7 +18,8 @@ Imports:
     tibble,
     rlang,
     fs (>= 1.3.0),
-    memoise
+    memoise,
+    httr
 License: Apache License 2.0 | file LICENSE
 URL: 
     http://centerforopenscience.github.io/osfr, 
