### 2019-06-02 0f9d527
Tweak download progress bars
