### 2019-06-02 ad01ef5
Tweak upload progress bar and warnings
