### 2019-06-02 b932dca
Progress bars are disabled by default
