### 2019-06-02 28a8b5f
Document progress argument
