### 2019-06-02 a631546
Update docs and bump dev version
