+++ NAMESPACE
@@ -51,6 +51,7 @@ importFrom(fs,path_ext_set)
 importFrom(fs,path_has_parent)
 importFrom(fs,path_rel)
 importFrom(fs,path_split)
+importFrom(httr,progress)
 importFrom(jsonlite,fromJSON)
 importFrom(memoise,forget)
 importFrom(memoise,memoise)
