### 2019-06-02 7234f85
Fix decompress argument for osf_downloads's generic
