### 2019-06-02 47d4d3d
Move progress bar text out of low-level endpoints
This was mainly down to accomodate `.wb_file_update` which doesn't
know the name of the file it's uploading.
