### 2019-06-02 5518eef
Replace unlink with fs delete functions
