### 2019-08-11 0abf496
Rewrite osf_downloads to implement `conflicts`
* .wb_downloads now invisibly returns TRUE
* Add check_local_dir()
* Add clean_path()
* Remove `decompress` argument
