+++ R/osf_download.R
@@ -12,13 +12,17 @@
 #' downloaded file. Note, the directory portion of `path` must reference an
 #' existing directory.
 #'
+#' OSF directories are downloaded as zip files and
+#'
 #' @param x An [`osf_tbl_file`] containing a single file or directory.
 #' @param path Path pointing to a local directory where the downloaded files
 #'   will be saved. Default is to use the current working directory.
-#' @param overwrite Logical, if the local path already exists should it be
-#'   replaced with the downloaded file?
-#' @param decompress Logical, should downloaded directories be automatically
-#'   unzipped? Default is `TRUE`.
+#' @param conflicts This determines what happens when a file with the same name
+#'   exists at the specified destination. Can be one of the following:
+#'   * `"error"` (the default): throw an error and abort the file transfer operation.
+#'   * `"skip"`: skip the conflicting file(s) and continue transferring the
+#'     remaining files.
+#'   * `"overwrite"`: replace the existing file with the transferred copy.
 #' @template progress
 #' @template verbose
 #'
@@ -38,13 +42,13 @@
 #' * [`osf_ls_files()`] for listing files and directories on OSF.
 #'
 #' @export
-#' @importFrom fs path_ext_set path_ext_remove dir_exists
+#' @importFrom fs path_ext_set path_ext_remove
 
 osf_download <-
   function(x,
            path = NULL,
-           overwrite = FALSE,
-           decompress = TRUE,
+           recurse = FALSE,
+           conflicts = "error",
            verbose = FALSE,
            progress = FALSE) {
   UseMethod("osf_download")
@@ -54,51 +58,187 @@ osf_download <-
 osf_download.osf_tbl_file <-
   function(x,
            path = NULL,
-           overwrite = FALSE,
-           decompress = TRUE,
+           recurse = FALSE,
+           conflicts = "error",
            verbose = FALSE,
            progress = FALSE) {
 
-  if (is.null(path)) {
-    path <- x$name
-  } else {
-    if (!fs::dir_exists(path))
-      abort("`path` must point to an existing local directory.")
-    path <- file.path(path, x$name)
+  path <- check_local_dir(path)
+
+  # convert recurse to number of recursion levels
+  if (is.logical(recurse)) {
+    recurse <- ifelse(isTRUE(recurse), Inf, 1)
   }
 
-  type <- ifelse(is_osf_dir(x), "folder", "file")
+  requests <- split(x, ifelse(is_osf_dir(x), "folder", "file"))
+  results <- vector(mode = "list")
 
-  path <- ifelse(
-    type == "folder",
-    as.character(fs::path_ext_set(path, "zip")),
-    path
+  if (!is.null(requests$folder)) {
+    n <- nrow(requests$folder)
+    results$folder <- Map(
+      f =  download_dir,
+      x = split(requests$folder, seq_len(n)),
+      path = path,
+      conflicts = conflicts,
+      recurse = recurse,
+      progress = progress,
+      verbose = verbose
+    )
+  }
+
+  if (!is.null(requests$file)) {
+    n <- nrow(requests$file)
+    results$file <- Map(
+      f =  download_file,
+      x = split(requests$file, seq_len(n)),
+      path = path,
+      conflicts = conflicts,
+      progress = progress,
+      verbose = verbose
+    )
+  }
+
+  # bind results within each type before returning merged file/folder slots
+  results <- lapply(results, do.call, what = "rbind")
+  do.call("rbind", results)
+}
+
+
+#' Internal download functions
+#'
+#' This is a non-vectorized function that downloads a single file. It
[…]
