### 2019-08-11 03371d4
Add new nested directory to test project
