### 2019-08-11 f9f6f4c
Refactor tests for new osf_download
