### 2019-08-17 675eb37
Better messages when skipping/downloading files
