### 2019-08-17 aa1e2ea
Remove internal unzip_files()
