### 2019-08-18 e6df770
Update docs for osf_downloads
