+++ R/osf_download.R
@@ -1,18 +1,22 @@
 #' Download files and directories from OSF
 #'
 #' @description
-#' Files stored on OSF can be downloaded locally by providing an
-#' [`osf_tbl_file`] that contains the OSF file of interest. If the
-#' [`osf_tbl_file`] contains a directory, a zip file containing the
-#' directory's contents will be downloaded.
+#' Files stored on OSF can be downloaded locally by passing an [`osf_tbl_file`]
+#' that contains the files and folders of interest. Use `path` to specify
+#' *where* the files should be downloaded, otherwise they are downloaded to
+#' your working directory by default.
 #'
-#' By default files are downloaded to your current working directory with the
-#' same filename used on OSF. The `path` argument can be used to specify a
-#' different destination *and*, optionally,  a different filename for the
-#' downloaded file. Note, the directory portion of `path` must reference an
-#' existing directory.
 #'
-#' OSF directories are downloaded as zip files and
+#' @section Implementation details:
+#' Directories are always downloaded from OSF as zip files that contain its
+#' entire contents. The logic for handling conflicts and recursion is
+#' implemented locally, acting on these files in a temporary location and
+#' copying them to `path` as needed. This creates a *gotcha* if you're
+#' downloading directories with large files and assuming that setting `conflicts
+#' = "skip"` and/or limiting recursion will reduce the number of files you're
+#' downloading. In such a case, a better strategy would be to use
+#' `osf_ls_files()` to list the contents of the directory and pass that output
+#' to `osf_download()`.
 #'
 #' @param x An [`osf_tbl_file`] containing a single file or directory.
 #' @param path Path pointing to a local directory where the downloaded files
@@ -23,11 +27,16 @@
 #'   * `"skip"`: skip the conflicting file(s) and continue transferring the
 #'     remaining files.
 #'   * `"overwrite"`: replace the existing file with the transferred copy.
+#' @param recurse Applies only to OSF directories. If `TRUE`, a directory is
+#'   fully recursed and all nested files and subdirectories are downloaded.
+#'   Alternatively, a positive number will determine the number of levels
+#'   to recurse.
 #' @template progress
 #' @template verbose
 #'
-#' @return The [`osf_tbl_file`] input with a new column, `"local_path"`,
-#'   containing the downloaded file's path.
+#' @return The same [`osf_tbl_file`] passed to `x` with a new column,
+#'   `"local_path"`, containing paths to the local files.
+#'
 #' @examples
 #' \dontrun{
 #' # download a single file
