### 2019-08-18 8ec4147
Make sync warning a template
