### 2019-08-19 a390ac2
Directories with a dot prefix can be created
