+++ tests/testthat/test-directories.R
@@ -82,3 +82,10 @@ test_that("'path' isn't confused by dir names with a shared substring (#95)", {
   d3 <- osf_mkdir(p1, path = "dir")
   expect_silent(osf_ls_files(p1, path = "dir"))
 })
+
+test_that("directory names can start with a dot", {
+  dotdir <- osf_mkdir(p1, path = ".dir")
+  expect_s3_class(dotdir, "osf_tbl_file")
+  expect_equal(dotdir$name, ".dir")
+})
+
