### 2019-08-19 f7eea79
Rename clean_path() -> clean_local_path()
