### 2019-08-19 #279 @aaronwolen
Hi @maelle, getting very close indeed.

Multi-file uploading and downloading now works fine ✅.

However, I'm still tweaking some of the logic for handling file conflicts. In my original submission `osf_upload()`/`osf_download()` had only a binary `overwrite` argument, which just doesn't scale to multiple files, especially if you want to do something like upload (and later update) an entire project directory.

My solution was to drop `overwrite` in favor of a new `conflicts` argument that allows you to optionally `"overwrite"` or `"skip"` conflicting files (either uploading or downloading), without stopping the entire transfer.

So this has involved quite a bit of refactoring and additional testing.

The changes are all still in a feature branch but I'm planning to merge to the dev branch soon and put out a call on the Slack channel for some brave beta testers to kick the tires.
