### 2019-08-20 acf27d0
Add clean_osf_path to strip reserved filenames from paths
