+++ R/osf_mkdir.R
@@ -74,7 +74,7 @@ osf_mkdir.osf_tbl_file <- function(x, path, verbose = FALSE) {
 #' determines what happens if an intermediate directory does not exist.
 #'
 #' @param x An `osf_tbl_node` or an `osf_tbl_file` with a directory.
-#' @param path A path of directories.
+#' @param path A scalar vector containing a path of directories.
 #' @param missing_action Either `"error"` or `"create"` to create the missing
 #'   directory.
 #' @importFrom fs path_rel
@@ -82,10 +82,13 @@ osf_mkdir.osf_tbl_file <- function(x, path, verbose = FALSE) {
 recurse_path <- function(x, path, missing_action = "error", verbose = FALSE) {
   missing_action <- match.arg(missing_action, c("error", "create"))
 
-  # skip leading slashes or dots so `path_root` is a valid folder name
-  path_dirs <- fs::path_split(path)[[1]]
-  valid_names <- grepl(sprintf("^[^\\%s]", .Platform$file.sep), path_dirs)
-  path_root <- path_dirs[valid_names][1]
+  stopifnot(rlang::is_scalar_character(path))
+
+  path_dirs <- fs::path_split(clean_osf_path(path))[[1]]
+  path_root <- path_dirs[1]
+
+  # return the original destination
+  if (path_root == ".") return(x)
 
   # ensure the retrieved directory and path_root have the same name
   root_dir <- osf_ls_files(x, type = "folder", pattern = path_root)
