### 2019-08-20 dc11a07
Refactor osf_upload()
- No longer using fs::dir_map
- Adds an internal upload_dir() to better mirror osf_download
