+++ tests/testthat/test-uploading-multiple-files.R
@@ -7,7 +7,7 @@ setup({
   # copy directory for testing multifile uploads to pwd
   multidir <<- fs::dir_copy(
     file.path(rprojroot::find_testthat_root_file(), "test-files", "uploads"),
-    "."
+    ".osfr-tests"
   )
 
   if (has_pat()) {
@@ -37,46 +37,85 @@ test_that("multiple files can be uploaded", {
 test_that("a directory can be uploaded", {
   skip_if_no_pat()
 
-  indir <- file.path(multidir, "subdir1", "subdir1_1")
-  out <- osf_upload(p1, indir)
+  c1_dir <- osf_create_component(p1, "dir-upload")
+  out <- osf_upload(c1_dir, multidir)
   expect_s3_class(out, "osf_tbl_file")
-  expect_equal(out$name, basename(indir))
+  expect_match(out$name, multidir)
+
+  # files within the directory were uploaded
+  expect_equal(
+    osf_ls_files(out, type = "file")$name,
+    basename(fs::dir_ls(multidir, type = "file"))
+  )
+
+  # subdirectories within the directory were created
+  subdirs <- basename(fs::dir_ls(multidir, type = "directory"))
+  expect_equal(
+    osf_ls_files(out, type = "folder")$name,
+    subdirs
+  )
+
+  # subdirectories are empty with recurse=FALSE
+  subdir1 <- osf_ls_files(out, path = subdirs[1])
+  expect_equal(nrow(subdir1), 0)
+})
+
+test_that("a subdirectory can be uploaded", {
+  skip_if_no_pat()
+
+  # upload the subdirectory with no children
+  c2_subdir <- osf_create_component(p1, "subdir-upload")
+
+  indir <- file.path(multidir, "subdir1", "subdir1_1")
+  out <- osf_upload(c2_subdir, indir)
+
+  expect_s3_class(out, "osf_tbl_file")
+  expect_match(out$name, basename(indir))
+
+  # verify leaf directory was created in the project root
+  expect_match(
+    osf_ls_files(c2_subdir, type = "folder")$name,
+    basename(indir)
+  )
 
   # verify files within the directory were uploaded
   expect_equal(osf_ls_files(out)$name, list.files(indir))
 })
 
-test_that("uploading does not recurse by default", {
+
+test_that("recurse argument respects specified levels", {
   skip_if_no_pat()
 
-  c1 <- osf_create_component(p1, "recurse = off")
-  out <- osf_upload(c1, multidir)
-  expect_equal(out$name, basename(multidir))
+  c3_recurse <- osf_create_component(p1, "recurse=1")
+  out <- osf_upload(c3_recurse, path = multidir, recurse = 1, verbose = TRUE)
 
-  # verify files and subdirectories were uploaded/created
-  expect_equal(osf_ls_files(out)$name, list.files(multidir))
+  sd1_files <- osf_ls_files(out, path = "subdir1")
+  expect_equal(
+    sd1_files$name,
+    dir(file.path(multidir, "subdir1"))
+  )
 
-  # but subdirectories should be empty
-  subdir1 <- osf_ls_files(out, path = "subdir1")
-  expect_equal(nrow(subdir1), 0)
-  subdir2 <- osf_ls_files(out, path = "subdir2")
-  expect_equal(nrow(subdir2), 0)
+  sd2_files <- osf_ls_files(out, path = "subdir2")
+  expect_equal(
+    sd2_files$name,
+    dir(file.path(multidir, "subdir2"))
+  )
+
+  sd1_1_files <- osf_ls_files(out, path = "subdir1/subdir1_1")
[…]
