### 2019-08-20 8b8ab40
Add test for uploading from parent directories
