+++ NAMESPACE
@@ -42,13 +42,14 @@ importFrom(crul,HttpClient)
 importFrom(crul,upload)
 importFrom(crul,url_parse)
 importFrom(fs,dir_exists)
-importFrom(fs,dir_walk)
 importFrom(fs,file_exists)
 importFrom(fs,file_info)
 importFrom(fs,is_dir)
+importFrom(fs,path_common)
 importFrom(fs,path_ext_remove)
 importFrom(fs,path_ext_set)
 importFrom(fs,path_has_parent)
+importFrom(fs,path_join)
 importFrom(fs,path_rel)
 importFrom(fs,path_split)
 importFrom(httr,progress)
