### 2019-08-20 #279 @cboettig
Thanks @aaronwolen ! This sounds great.  standing by for the ping when you're ready for those beta testers.
