### 2019-08-25 6d3d59d
Small re-org of upload code
