### 2019-08-25 321e9a0
Add internal function for finding existing OSF files and folders
