### 2019-08-26 062a10e
Implement `conflicts` argument for osf_upload
