+++ R/osf_upload.R
@@ -10,11 +10,7 @@
 #' @param recurse If `TRUE`, fully recurse directories included in `path`. You
 #'   can also control the number of levels to recurse by specifying a positive
 #'   number.
-#' @param overwrite Logical, overwrite an existing file with the same name
-#'   (default `FALSE`)? If `TRUE`, OSF will automatically update the file and
-#'   record the previous version. If `FALSE`, a warning will be issued that the
-#'   local file was *not* uploaded and the *existing* version of the file on OSF
-#'   is returned.
+#' @template conflicts
 #' @template progress
 #' @template verbose
 #'
@@ -96,7 +92,7 @@ osf_upload <-
   function(x,
            path,
            recurse = FALSE,
-           overwrite = FALSE,
+           conflicts = "error",
            progress = FALSE,
            verbose = FALSE) {
   UseMethod("osf_upload")
@@ -107,13 +103,13 @@ osf_upload.osf_tbl_node <-
   function(x,
            path,
            recurse = FALSE,
-           overwrite = FALSE,
+           conflicts = "error",
            progress = FALSE,
            verbose = FALSE) {
 
   path <- check_files(path)
   x <- make_single(x)
-  .osf_upload(x, path, recurse, overwrite, progress, verbose)
+  .osf_upload(x, path, recurse, conflicts, progress, verbose)
 }
 
 #' @export
@@ -121,7 +117,7 @@ osf_upload.osf_tbl_file <-
   function(x,
            path,
            recurse = FALSE,
-           overwrite = FALSE,
+           conflicts = "error",
            progress = FALSE,
            verbose = FALSE) {
 
@@ -136,7 +132,7 @@ osf_upload.osf_tbl_file <-
     ))
   }
 
-  .osf_upload(x, path, recurse, overwrite, progress, verbose)
+  .osf_upload(x, path, recurse, conflicts, progress, verbose)
 }
 
 
@@ -148,7 +144,7 @@ osf_upload.osf_tbl_file <-
 #' @importFrom fs is_dir file_info
 #' @noRd
 
-.osf_upload <- function(dest, path, recurse, overwrite, progress, verbose) {
+.osf_upload <- function(dest, path, recurse, conflicts, progress, verbose) {
 
   # split into top-level files and folders
   path_by <- split(path, fs::file_info(path)$type, drop = TRUE)
@@ -159,7 +155,7 @@ osf_upload.osf_tbl_file <-
     results$file <- purrr::map(path_by$file,
       .upload_file,
       dest = dest,
-      overwrite = overwrite,
+      conflicts = conflicts,
       progress = progress,
       verbose = verbose
     )
@@ -170,7 +166,7 @@ osf_upload.osf_tbl_file <-
     results$directory <- purrr::map(path_by$directory,
       .upload_dir,
       dest = dest,
-      overwrite = overwrite,
+      conflicts = conflicts,
       recurse = recurse,
       progress = progress,
       verbose = verbose
@@ -185,7 +181,7 @@ osf_upload.osf_tbl_file <-
 #' @param path a scalar character vector containing a single directory path
 #' @return An `osf_tbl_file` containing a single row for `path`'s leaf directory
 #' @noRd
-.upload_dir <- function(dest, path, overwrite, recurse, progress, verbose) {
+.upload_dir <- function(dest, path, conflicts, recurse, progress, verbose) {
   stopifnot(rlang::is_scalar_character(path))
 
   # memoise osf directory retrieval to avoid subsequent API calls for every
@@ -220,7 +216,7 @@ osf_upload.osf_tbl_file <-
       f = .upload_file,
       dest = targets$file$destination,
       path = targets$file$local,
-      overwrite = overwrite,
+      conflicts = conflicts,
       progress = progress,
[…]
