### 2019-08-31 5fe0583
Tweak osf_upload docs
