### 2019-08-31 0fb4c0e
Move dev utilities to separate file
