### 2019-08-31 397b407
Add internal map_rbind()
