### 2019-08-31 2f06a51
Another revision to osf_upload
- replaces .upload_dir with .upload_manifest
