+++ R/utils-uploads.R
@@ -1,0 +1,28 @@
+#' Create manifest of local files/directories to upload
+#'
+#' @param paths a vector of file paths for local files and directories
+#' @param recurse determines whether directories in `paths` are recursed
+#' @return data.frame of file metadata generated by `fs::file_info`, with the
+#'   the following additional columns: `remote_path` and `remote_dir`
+#' @noRd
+.upload_manifest <- function(paths, recurse) {
+  stopifnot(rlang::is_scalar_vector(filepath))
+
+  if (fs::is_dir(filepath)) {
+    targets <- fs::dir_ls(filepath, type = c("file", "directory"), recurse = recurse)
+  } else {
+    targets <- filepath
+  }
+
+  inventory <- fs::file_info(targets)
+
+  # the leaf directory of `filepath` will be the base destination for uploads
+  inventory$remote_path <- fs::path_rel(inventory$path, dirname(filepath))
+  inventory$remote_dir <- ifelse(
+    inventory$type == "file",
+    dirname(inventory$remote_path),
+    identity(inventory$remote_path)
+  )
+
+  inventory
+}
