### 2019-08-31 b7df8fb
Bump dev version and update docs
