### 2019-09-01 cf6f2db
Non-verbose messages for osf_upload
