+++ R/utils-uploads.R
@@ -1,23 +1,23 @@
 #' Create manifest of local files/directories to upload
 #'
-#' @param paths a vector of file paths for local files and directories
-#' @param recurse determines whether directories in `paths` are recursed
+#' @param path a vector of file paths for local files and directories
+#' @param recurse determines whether directories in `path` are recursed
 #' @return data.frame of file metadata generated by `fs::file_info`, with the
 #'   the following additional columns: `remote_path` and `remote_dir`
 #' @noRd
-.upload_manifest <- function(paths, recurse) {
-  stopifnot(rlang::is_scalar_vector(filepath))
+.upload_manifest <- function(path, recurse) {
+  stopifnot(rlang::is_scalar_vector(path))
 
-  if (fs::is_dir(filepath)) {
-    targets <- fs::dir_ls(filepath, type = c("file", "directory"), recurse = recurse)
+  if (fs::is_dir(path)) {
+    targets <- fs::dir_ls(path, type = c("file", "directory"), recurse = recurse)
   } else {
-    targets <- filepath
+    targets <- path
   }
 
   inventory <- fs::file_info(targets)
 
-  # the leaf directory of `filepath` will be the base destination for uploads
-  inventory$remote_path <- fs::path_rel(inventory$path, dirname(filepath))
+  # the leaf directory of `path` will be the base destination for uploads
+  inventory$remote_path <- fs::path_rel(inventory$path, dirname(path))
   inventory$remote_dir <- ifelse(
     inventory$type == "file",
     dirname(inventory$remote_path),
