### 2019-09-01 ae7ab95
Skipped file messages only when verbose is TRUE
