### 2019-09-01 e03fcce
Test that d/l'ed files can overwrite conflicts
