### 2019-09-01 a0eeb67
Fix downloading when conflicts='overwrite'
