### 2019-09-01 cdec997
Minor cleanup of osf_download/osf_upload
- Make internal variables more consistent
- Move ancilliary funtctions to separate utils files
