### 2019-09-01 e473b2b
Use `conflicts` argument in vignette (#103)
