### 2019-09-01 e5996f4
Improve NEWS and bump dev version
