### 2019-09-01 a2506ea
Simplify travis config
