+++ .travis.yml
@@ -2,8 +2,17 @@ language: r
 sudo: false
 cache: packages
 
-r_build_args: --no-build-vignettes
-r_check_args: --ignore-vignettes
+
+r_packages:
+  - devtools
+
+install:
+- R -e 'install.packages("remotes")'
+- R -e 'remotes::install_deps(dep = T)'
+
+script:
+- R CMD build . --no-build-vignettes
+- travis_wait 20 R CMD check *tar.gz --ignore-vignettes
 
 matrix:
   include:
