### 2019-09-02 cbc141f
Update links to OSF help docs
