### 2019-09-02 4f6a67b
Update links to OSF help docs
