### 2019-09-02 0d97654
Export osf_refresh()
