+++ R/osf_refresh.R
@@ -3,7 +3,8 @@
 #' Use `osf_refresh()` to update one or more entities in an [`osf_tbl()`] with
 #' the latest information from OSF.
 #'
-#' @param x an [`osf_tbl`]
+#' @param x an [`osf_tbl`].
+#' @export
 
 osf_refresh <- function(x) {
   stopifnot(inherits(x, "osf_tbl"))
@@ -14,8 +15,7 @@ osf_refresh <- function(x) {
                      osf_tbl_user = osf_retrieve_user
   )
 
-  res <- lapply(as_id(x), retrieve)
-  refreshed <- do.call("rbind", res)
+  refreshed <- map_rbind(retrieve, as_id(x))
 
   # update x to preserve any additional variables
   x[colnames(refreshed)] <- refreshed
