+++ NAMESPACE
@@ -33,6 +33,7 @@ export(osf_ls_nodes)
 export(osf_mkdir)
 export(osf_mv)
 export(osf_open)
+export(osf_refresh)
 export(osf_retrieve_file)
 export(osf_retrieve_node)
 export(osf_retrieve_user)
