### 2019-09-02 7b02df9
An OSF dir is created when uploading an empty dir
