### 2019-10-10 b4cd952
Minor changes to NEWS
