### 2019-10-10 c3544bb
Bump version and update docs
