### 2019-10-10 b54b0c6
Bump version and update docs
