### 2019-10-10 #279 @aaronwolen
You know what? You're never going to believe this... but I think it's ready for testing. 😮

So @cboettig and @HeidiSeibold, if you're still around after all this time, I'd love to get some feedback on the new upload/download implementations. In particular, it'd be great to know if the docs are clear and whether the new behavior of `osf_upload()`/`osf_download()` matches your expectations.

If you're willing to give it a go, `v0.2.5` (which I'm calling the beta) is available on the `dev` branch.

```
remotes::install_github("centerforopenscience/osfr", ref = "dev")
```
