### 2019-10-11 4aa47e5
Separate file for osf_tbl utils
