### 2019-10-11 a9fa0c7
Set lintr minimum version
