### 2019-10-11 017b402
Address (or punt) minor lint issues
