+++ R/zzz.R
@@ -1,15 +1,15 @@
-.onLoad <- function(libname, pkgname) {
+.onLoad <- function(libname, pkgname) { # nolint
 
   # record personal access token
-  env.pat <- Sys.getenv("OSF_PAT")
-  if (nzchar(env.pat)) options(osfr.pat = env.pat)
+  env_pat <- Sys.getenv("OSF_PAT")
+  if (nzchar(env_pat)) options(osfr.pat = env_pat)
 
   # setup logger
-  env.log <- Sys.getenv("OSF_LOG")
-  if (nzchar(env.log)) {
+  env_log <- Sys.getenv("OSF_LOG")
+  if (nzchar(env_log)) {
     if (requireNamespace("logger", quietly = TRUE)) {
-      options(osfr.log = env.log)
-      logger::log_appender(logger::appender_file(env.log), namespace = "osfr")
+      options(osfr.log = env_log)
+      logger::log_appender(logger::appender_file(env_log), namespace = "osfr")
       logger::log_formatter(logger::formatter_sprintf, namespace = "osfr")
     } else {
       warn("The logger package must installed to enable logging")
@@ -29,7 +29,7 @@
   invisible()
 }
 
-.onAttach <- function(libname, pkgname) {
+.onAttach <- function(libname, pkgname) { # nolint
   if (!is.null(getOption("osfr.pat"))) {
     packageStartupMessage("Automatically registered OSF personal access token")
   }
