### 2019-10-11 14dcda7
Refactor internal build_move_request()
Avoid internal s3 method to satisfy lintr
