### 2019-10-11 965b9fb
Bump dev version and update docs
