### 2019-10-11 #279 @aaronwolen
Not at all, @HeidiSeibold. I'd appreciate your feedback whenever you get a chance.

Since these changes enable osfr to handle (potentially) lots of files at once, I really want to  test them out as thoroughly as possible before opening them up to a wider audience.
