### 2019-11-04 #279 @cboettig
Quite happy with the new interface on my end!   Well done!  having streamlined multifile upload is really quite nice.  Would love to see this on CRAN soon??

Another question came up as I tried recruiting some more testers -- the process of "publishing" data with a DOI is a bit unclear from the API (or really from my quick glance at the OSF documentation).  I'm pretty sure this is possible, but could you include a quick example in a vignette somewhere that illustrates how to publish data under a DOI?  (Ideally I think this could be done with either a two-step process, where you can first 'reserve' the DOI and add it to the repo docs/ internal metadata before cutting the release, or a one-shot publish with a DOI).  Happy to bounce ideas back and forth further on this.
