### 2019-11-13 CenterForOpenScience/osfr#107 @aaronwolen
Generating external IDs
The API provides endpoints for generating and listing external IDs (currently limited to DOIs). We could implement something like

```r
osf_generate_id(proj, type = "doi")
```

which would return a modified version of `proj` with a new `doi` column.

* The generic name provides for the possibility of being able to generate other IDs
* *generate* is annoyingly long... might need a different verb
* Should it verify a node is public? Is that a requirement?

Would also need `osf_ls_ids()` to list previously generated IDs since they aren't returned by default.

Comments/suggestions, @cboettig?
