### 2019-11-13 #279 @aaronwolen
Hey @cboettig, somehow missed your response! Thanks for the positive feedback.

Good idea to include some info about generating DOIs in the vignette, since that is one of OSF's killer features. I also like the idea of being able to do it directly with osfr, I created [an issue](https://github.com/CenterForOpenScience/osfr/issues/107) to explore adding this feature—although I'm going to hold off adding new features until this thing actually lands on CRAN 😄.

Protocol is to hold off on submitting until the review has concluded, correct?
