### 2019-11-18 7b0dfcd
Stricter assertions for get_parent_id
