### 2019-11-18 40121a3
Replace purrr::pluck with purrr::chuck
