+++ R/osfr-package.R
@@ -7,7 +7,7 @@
 #' @importFrom crul HttpClient url_parse
 #' @importFrom httr progress
 #' @importFrom stringi stri_split_fixed
-#' @importFrom purrr map map_chr imap_chr pluck set_names transpose %||%
+#' @importFrom purrr map map_chr imap_chr chuck set_names transpose %||%
 #' @importFrom tibble tibble
 #' @importFrom rlang abort warn is_empty is_scalar_character
 #' @importFrom utils head tail modifyList browseURL unzip
