### 2019-11-18 be822de
Add wb2osf()
