### 2019-11-18 49c8a41
New osf_upload strategy
- We now check for conflicting files on OSF before uploading, rather
  than uploading and looking for 409 errors
- Much of the logic for identifying remote destinations and identifying
  conflicting files has been subsumed in the upload file manifest
  utilities
- The internal upload function has been broken into .upload_new_file and
  .update_existing file, both of which return processed waterbutler
  responses. We now defer converting these entities to osf_tbls so that
  we can filter for files/dirs directly specified in 'path', reducing
  the number of file retrieval
[…]
