+++ R/utils-uploads.R
@@ -1,34 +1,3 @@
-#' Create manifest of local files/directories to upload
-#'
-#' @param path a vector of file paths for local files and directories
-#' @param recurse determines whether directories in `path` are recursed
-#' @return data.frame of file metadata generated by `fs::file_info`, with the
-#'   the following additional columns: `remote_path` and `remote_dir`
-#' @noRd
-.upload_manifest <- function(path, recurse) {
-  stopifnot(rlang::is_scalar_vector(path))
-
-  if (fs::is_dir(path)) {
-    targets <- fs::dir_ls(path, type = c("file", "directory"), recurse = recurse)
-  }
-
-  if (fs::is_file(path) || rlang::is_empty(targets)) {
-    targets <- path
-  }
-
-  inventory <- fs::file_info(targets)
-
-  # the leaf directory of `path` will be the base destination for uploads
-  inventory$remote_path <- fs::path_rel(inventory$path, dirname(path))
-  inventory$remote_dir <- ifelse(
-    inventory$type == "file",
-    dirname(inventory$remote_path),
-    identity(inventory$remote_path)
-  )
-
-  inventory
-}
-
 #' File transfer messages
 #'
 #' Generate messages about file conflicts while uploading or downloading.
@@ -59,15 +28,65 @@ warn_ul_conflict <- function(filename) {
 }
 
 
-#' Succinct message indicating number of files uploaded or skipped
-#' @param x an `osf_tbl_file` with the additional `.uploaded` column
+#' Upload new file
+#'
+#' @param dest OSF node or directory upload destination
+#' @param path scalar character vector with the path of the file to be uploaded
+#' @return A list object containing the processed waterbutler json response
 #' @noRd
-report_ul_activity <- function(x) {
-  if (any(x$.uploaded))
-    message(sprintf("Uploaded %i file(s) to OSF", sum(x$.uploaded)))
-  if (any(!x$.uploaded))
-    message(sprintf(
-      "Skipped %i file(s) to avoid overwriting OSF copies",
-      sum(!x$.uploaded)
-    ))
+
+.upload_new_file <- function(dest, path, progress, verbose) {
+  stopifnot(inherits(dest, c("osf_tbl_node", "osf_tbl_file")))
+  stopifnot(rlang::is_scalar_character(path))
+
+  # force the uploaded filename to match the local filename
+  filename <- basename(path)
+
+  # set arguments depending on whether destination is a directory or node
+  upload_args <- list(
+    name = filename,
+    body = crul::upload(path),
+    progress = progress
+  )
+
+  if (inherits(dest, "osf_tbl_node")) {
+    upload_args$id <- as_id(dest)
+  } else {
+    upload_args$id <- get_parent_id(dest)
+    upload_args$fid <- as_id(dest)
+  }
+
+  if (progress) cat(sprintf("Uploading %s\n", filename))
+  res <- do.call(".wb_file_upload", upload_args)
+
+  raise_error(res)
+  if (verbose) message(sprintf("Uploaded new file '%s' to OSF", filename))
+  res
+}
+
+#' Update existing file
+#' @param dest An osf_tbl_file with the remote file to be overwritten
+#' @param path scalar character vector with the path of the file to be uploaded
+#' @return A list object containing the processed waterbutler json response
+#' @noRd
+
+.update_existing_file <- function(dest, path, progress, verbose) {
+  stopifnot(inherits(dest, "osf_tbl_file"))
+  stopifnot(rlang::is_scalar_character(path))
+
+  # force the uploaded filename to match the local filename
+  filename <- basename(path)
+
+  if (progress) cat(sprintf("Uploading file '%s'\n", filename))
+
[…]
