### 2019-11-23 e77f768
Add permissions/doi links to vignette
