### 2019-11-23 CenterForOpenScience/osfr#107 @aaronwolen
Also like the following function names for making entities public/private:
* `osf_publicize()`
* `osf_privatize()`

They'll both need `recurse` arguments to control whether lower-level components are affected.
