### 2019-11-24 93c3964
travis print log on failure
