+++ DESCRIPTION
@@ -37,5 +37,5 @@ Suggests:
     covr
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 6.1.1
+RoxygenNote: 7.0.1
 VignetteBuilder: knitr
