### 2019-11-24 de03333
Skip subclass determination on empty osf_tbls
Not useful since determination requires a non-empty meta column.
This also avoids errors when attempting to rbind an empty osf_tbl
with a non-empty osf_tbl_file.
