### 2019-11-24 0e9d5a3
.wb_file_move now returns an osf_tbl
