### 2019-12-01 053fc55
Remove hardcoded progress=TRUE
