### 2019-12-01 6a3a99e
Improve upload messaging
