### 2019-12-01 e60d881
Bumb version and update news [skip ci]
