### 2019-12-03 8e6d732
Tweak travis builds
