+++ .travis.yml
@@ -2,10 +2,6 @@ language: r
 sudo: false
 cache: packages
 
-
-r_packages:
-  - devtools
-
 install:
 - R -e 'install.packages("remotes")'
 - R -e 'remotes::install_deps(dep = T)'
@@ -14,21 +10,18 @@ script:
 - R CMD build . --no-build-vignettes
 - travis_wait 20 R CMD check *tar.gz --ignore-vignettes
 
-jobs:
+matrix:
   include:
   - r: oldrel
   - r: devel
   - r: release
     after_success:
     - Rscript -e 'covr::codecov()'
-    before_cache:
-    - Rscript -e 'remotes::install_cran("pkgdown")'
     after_failure:
     - cat osfr.log
-
+    before_cache:
+    - Rscript -e 'remotes::install_cran("pkgdown")'
     deploy:
-      on:
-        branch: master
       provider: script
       script: Rscript -e 'pkgdown::deploy_site_github(verbose = TRUE)'
       skip_cleanup: true
