+++ .travis.yml
@@ -22,6 +22,8 @@ matrix:
     before_cache:
     - Rscript -e 'remotes::install_cran("pkgdown")'
     deploy:
+      on:
+        branch: fix-pkgdown
       provider: script
       script: Rscript -e 'pkgdown::deploy_site_github(verbose = TRUE)'
       skip_cleanup: true
