### 2019-12-04 1688a93
Update travis config to dpl v2
