### 2019-12-04 a0a4391
Temporarily skip tests on travis
