### 2019-12-04 6a07ee4
Remove unnecessary travis steps
