+++ .travis.yml
@@ -11,15 +11,16 @@ jobs:
   - r: oldrel
   - r: devel
   - r: release
-    after_success:
-    - Rscript -e 'covr::codecov()'
+    # after_success:
+    # - Rscript -e 'covr::codecov()'
     after_failure:
     - cat osfr.log
     before_cache:
-    - Rscript -e 'remotes::install_cran("pkgdown")'
+    - Rscript -e 'dir(recursive = TRUE)' && Rscript -e 'remotes::install_cran("pkgdown")'
     deploy:
       on:
         branch: fix-pkgdown
       provider: script
-      script: Rscript -e 'pkgdown::deploy_site_github(verbose = TRUE)'
+      script: Rscript -e 'dir(recursive = TRUE)' && Rscript -e 'pkgdown::deploy_site_github(verbose = TRUE)'
       cleanup: false
+
