### 2019-12-04 5ebb756
Export PKG_TARBALL variable in travis deploy
pkgdown was not finding the package tarball even though I verified
it was present prior to deployment
