### 2019-12-06 54509df
add copy function and edit example in cp/mv
