+++ man/osf_download.Rd
@@ -4,14 +4,8 @@
 \alias{osf_download}
 \title{Download files and directories from OSF}
 \usage{
-osf_download(
-  x,
-  path = NULL,
-  recurse = FALSE,
-  conflicts = "error",
-  verbose = FALSE,
-  progress = FALSE
-)
+osf_download(x, path = NULL, recurse = FALSE, conflicts = "error",
+  verbose = FALSE, progress = FALSE)
 }
 \arguments{
 \item{x}{An \code{\link{osf_tbl_file}} containing a single file or directory.}
@@ -67,7 +61,7 @@ to \code{osf_download()}.
 While \code{osf_download()} and \code{osf_upload()} can be used to conveniently shuttle
 files back and forth between OSF and your local machine, it's important to
 note that \strong{they are not file synchronization functions}. In contrast to
-something like \href{https://rsync.samba.org}{\code{rsync}},
+something like \href{https://rsync.samba.org}{rsync},
 \code{osf_download()}/\code{osf_upload()} do not take into account a file's contents or
 modification time. Whether you're uploading or downloading, if \code{overwrite = TRUE}, osfr will overwrite an existing file regradless of whether the
 existing file is the more recent copy. You have been warned.
