+++ NAMESPACE
@@ -4,6 +4,7 @@ S3method("[",osf_tbl)
 S3method(as_id,character)
 S3method(as_id,osf_id)
 S3method(as_id,osf_tbl)
+S3method(osf_cp,osf_tbl_file)
 S3method(osf_download,osf_tbl_file)
 S3method(osf_ls_files,osf_tbl_file)
 S3method(osf_ls_files,osf_tbl_node)
@@ -25,6 +26,7 @@ S3method(rbind,osf_tbl)
 export("%>%")
 export(as_id)
 export(osf_auth)
+export(osf_cp)
 export(osf_create_component)
 export(osf_create_project)
 export(osf_download)
