### 2019-12-06 04887b2
add copy function and edit example in cp/mv
