+++ .travis.yml
@@ -16,10 +16,10 @@ jobs:
     after_failure:
     - cat osfr.log
     before_deploy:
-    - Rscript -e 'dir(recursive = TRUE)' && Rscript -e 'remotes::install_cran("pkgdown")'
+    - Rscript -e 'remotes::install_cran("pkgdown")'
     deploy:
       on:
         branch: fix-pkgdown
       provider: script
-      script: Rscript -e 'dir(recursive = TRUE)' && Rscript -e 'pkgdown::deploy_site_github(verbose = TRUE)'
+      script: Rscript -e 'pkgdown::deploy_site_github(verbose = TRUE, tarball = dir(pattern = "tar.gz$"))'
       skip_cleanup: true
