### 2019-12-07 2b341d5
Tweak travis build
- sudo is deprecated
- os should be defined
- installing remotes and deps isn't necessary
