+++ .travis.yml
@@ -23,5 +23,5 @@ jobs:
       on:
         branch: master
       provider: script
-      script: Rscript -e 'pkgdown::deploy_site_github(verbose = TRUE)'
+      script: Rscript -e 'pkgdown::deploy_site_github(verbose = TRUE, tarball = dir(pattern = "tar.gz$"))'
       skip_cleanup: true
