### 2019-12-07 8455052
Bump version and update news
