### 2019-12-07 10e9505
Cleanup travis
