### 2019-12-07 745f07d
Include osfr log in artifacts
