### 2019-12-07 76cf3c9
Enable remote debugging for appveyor
