### 2019-12-07 320d979
Avoid dropping trailing slashes on Windows (closes #112)
