### 2019-12-07 afd6213
Add url_path() tests
