### 2019-12-08 0ca1979
Skip 2 tests when PAT is undefined
