### 2019-12-08 92a9cd2
Bump version and update news
