### 2019-12-10 ffdbf0f
add cp tests
