+++ tests/testthat/test-files.R
@@ -88,32 +88,40 @@ test_that("moving a parent directory to a child directory errors", {
 
 context("Copying files")
 
-test_that("a single file can be copied with appropriate suffix", {
+test_that("a file can copy to new directory", {
   skip_if_no_pat()
+  f1 <- osf_refresh(f1)
+  f_copy <- osf_cp(f1, d3)
+  expect_identical(get_parent_id(f_copy), get_parent_id(f1))
+  osf_rm(f_copy, check = FALSE)
+})
+
+test_that("a file cannot copy to same location with either overwrite option", {
+  skip_if_no_pat()
+  f1 <- osf_refresh(f1)
+  expect_error(
+    osf_cp(f1, p2, overwrite = FALSE),
+    "Cannot complete action:"
+  )
+
+  expect_error(
+    osf_cp(f1, p2, overwrite = TRUE),
+    "Unable to move or copy"
+  )
+})
+
+test_that("copy will respect overwrite values when copying to a new location", {
+  skip_if_no_pat()
+  f1 <- osf_refresh(f1)
+  f_copy <- osf_cp(f1, d3)
+  expect_error(
+    osf_cp(f_copy, d3, overwrite = FALSE),
+    "Cannot complete action:"
+  )
 
   f1 <- osf_refresh(f1)
-  f_copy <- osf_cp(f1)
-  expect_identical(get_parent_id(f_copy), get_parent_id(f1))
-  expect_identical(f_copy$name, paste0(fs::path_ext_remove(f1$name),  " (1).", fs::path_ext(f1$name)))
-})
-
-
-test_that("a directory can be copied with appropriate suffix", {
-  skip_if_no_pat()
-
-  #d3 <- osf_mkdir(p2, "d3")
-  d3_copy <- osf_cp(d3)
-  expect_identical(get_parent_id(d3_copy), get_parent_id(d3))
-  expect_identical(d3_copy$name, paste0(d3$name,  " (1)"))
-})
-
-
-test_that("a directory can be copied to a new node", {
-  skip_if_no_pat()
-
-  d3_copy2 <- osf_cp(d3, p1)
-  expect_identical(get_parent_id(d3), (p2$id))
-  expect_identical(get_parent_id(d3_copy2), (p1$id))
+  f_copy_2 <- osf_cp(f1, d3, overwrite = TRUE)
+  expect_identical(get_parent_id(f_copy_2), get_parent_id(f1))
 })
 
 
