### 2019-12-17 #279 @aaronwolen
All major changes are now available on the master branch and the paper should be done this week! 🤞
