### 2019-12-19 c466ca3
First pass at implementing vcr tests
