+++ tests/testthat/cassettes/test-directories.yml
@@ -1,0 +1,195 @@
+http_interactions:
+- request:
+    method: post
+    uri: https://api.test.osf.io/v2/nodes/
+    body:
+      encoding: ''
+      string: '{"data":{"type":"nodes","attributes":{"title":"osfr-test-directories","category":"project","description":"","public":false}}}'
+    headers:
+      Content-Type: application/json
+      Accept-Encoding: gzip, deflate
+      Accept: application/json, text/xml, application/xml, */*
+      User-Agent: osfr v0.2.6.2.9001
+      Authorization: Bearer <totallyrealpat>
+      Accept-Header: application/vnd.api+json;version=2.8
+  response:
+    status:
+      status_code: '201'
+      message: Created
+      explanation: Document created, URL follows
+    headers:
+      status: HTTP/1.1 201 Created
+      server: nginx
+      date: Wed, 08 Jan 2020 15:44:50 GMT
+      content-type: application/vnd.api+json
+      content-length: '3931'
+      vary: Accept
+      x-frame-options: SAMEORIGIN
+      allow: GET, POST, PUT, PATCH, DELETE, HEAD, OPTIONS
+      cache-control: no-cache, no-store, max-age=0, must-revalidate
+      expires: Mon, 01 Jan 1990 00:00:00 GMT
+      pragma: no-cache
+      via: 1.1 google
+      alt-svc: clear
+    body:
+      encoding: UTF-8
+      file: no
+      string: '{"data":{"relationships":{"logs":{"links":{"related":{"href":"https://api.test.osf.io/v2/nodes/q46yw/logs/","meta":{}}}},"contributors":{"links":{"related":{"href":"https://api.test.osf.io/v2/nodes/q46yw/contributors/","meta":{}}}},"citation":{"data":{"type":"nodes","id":"q46yw"},"links":{"related":{"href":"https://api.test.osf.io/v2/nodes/q46yw/citation/","meta":{}}}},"linked_registrations":{"links":{"self":{"href":"https://api.test.osf.io/v2/nodes/q46yw/relationships/linked_registrations/","meta":{}},"related":{"href":"https://api.test.osf.io/v2/nodes/q46yw/linked_registrations/","meta":{}}}},"children":{"links":{"related":{"href":"https://api.test.osf.io/v2/nodes/q46yw/children/","meta":{}}}},"view_only_links":{"links":{"related":{"href":"https://api.test.osf.io/v2/nodes/q46yw/view_only_links/","meta":{}}}},"comments":{"links":{"related":{"href":"https://api.test.osf.io/v2/nodes/q46yw/comments/?filter%5Btarget%5D=q46yw","meta":{}}}},"linked_by_registrations":{"links":{"related":{"href":"https://api.test.osf.io/v2/nodes/q46yw/linked_by_registrations/","meta":{}}}},"forks":{"links":{"related":{"href":"https://api.test.osf.io/v2/nodes/q46yw/forks/","meta":{}}}},"files":{"links":{"related":{"href":"https://api.test.osf.io/v2/nodes/q46yw/files/","meta":{}}}},"draft_registrations":{"links":{"related":{"href":"https://api.test.osf.io/v2/nodes/q46yw/draft_registrations/","meta":{}}}},"bibliographic_contributors":{"links":{"related":{"href":"https://api.test.osf.io/v2/nodes/q46yw/bibliographic_contributors/","meta":{}}}},"groups":{"links":{"related":{"href":"https://api.test.osf.io/v2/nodes/q46yw/groups/","meta":{}}}},"affiliated_institutions":{"links":{"self":{"href":"https://api.test.osf.io/v2/nodes/q46yw/relationships/institut
[…]
