### 2020-01-08 a0dc4d5
bump to dev version and update news
