### 2020-01-08 b3d9e33
bump to dev version and update news
