### 2020-01-08 b4afd2a
handle html symbols in node titles (fixes #117)
