### 2020-01-08 c19b34d
handle html symbols in node titles (fixes #117)
