### 2020-01-10 04a74e7
dry out cp/mv docs
