### 2020-01-10 237d968
dry out cp/mv docs
