### 2020-01-10 e5b4636
move internal file operations to a separate file
