### 2020-01-10 1968107
move internal file operations to a separate file
