### 2020-01-10 b0ca5b4
rename osf_rm()'s recursive argument to recurse
