### 2020-01-10 ce9f1d2
rename osf_rm()'s recursive argument to recurse
