### 2020-01-10 f737451
bump version and update docs
