### 2020-01-10 #279 @aaronwolen
Hi, @maelle! Just wanted to let you know the [paper](https://github.com/CenterForOpenScience/osfr/blob/master/inst/paper.md) has been pushed to the repo.
