### 2020-01-12 f5932ad
typos
ROpenSci -> rOpenSci and depply -> deeply
