### 2020-01-12 #279 @maelle
@HeidiSeibold I've just realized you didn't get a chance to approve the package (the remaining box in [your review](https://github.com/ropensci/software-review/issues/279#issuecomment-458882143)) -- it's fine if you don't have time to. If you don't answer within a week I'll do a few more checks myself.

@aaronwolen

* I noticed that the docs of ?osfr don't include more info about components/folders as discussed [in a comment of yours](https://github.com/ropensci/software-review/issues/279#issuecomment-462381709) and the following comment? The explanations in the paper are very clear. You could re-use an Rmd fragment in the README, vignette and also package level docs, see [this blog post about Rmd fragments](https://www.garrickadenbuie.com/blog/dry-vignette-and-readme/) and roxygen2 docs about Rmd use](https://roxygen2.r-lib.org/articles/rd.html#including-external--rmd-md-files).

* @cboettig had written "I would appreciate more guidance up-front on file size limits, project size limits, and permission management in the Getting Started Vignette, etc.". Has this been added?

* The paper looks fine (minus 2 typos in a PR) but where is paper.bib?
