+++ inst/paper.bib
@@ -1,0 +1,91 @@
+@Manual{crul,
+    title = {crul: HTTP Client},
+    author = {Scott Chamberlain},
+    year = {2019},
+    note = {R package version 0.9.0},
+    url = {https://CRAN.R-project.org/package=crul},
+  }
+
+@Manual{googledrive,
+  title = {googledrive: An Interface to Google Drive},
+  author = {{D'Agostino McGowan}, Lucy AND Bryan, Jennifer},
+  year = {2019},
+  note = {R package version 1.0.0},
+  url = {https://cran.r-project.org/package=googledrive},
+}
+
+@Manual{magrittr,
+    title = {magrittr: A Forward-Pipe Operator for R},
+    author = {Stefan Milton Bache and Hadley Wickham},
+    year = {2014},
+    note = {R package version 1.5},
+    url = {https://CRAN.R-project.org/package=magrittr},
+  }
+
+@article{Ram:2013de,
+  author = {Ram, Karthik},
+  title = {Git can facilitate greater reproducibility and increased transparency in science},
+  journal = {Source code for biology and medicine},
+  year = {2013},
+  volume = {8},
+  number = {1},
+  pages = {7},
+  month = feb
+}
+
+@article{Sandve:2013,
+  author = {Sandve, Geir Kjetil AND Nekrutenko, Anton AND Taylor, James AND Hovig, Eivind},
+  journal = {PLOS Computational Biology},
+  publisher = {Public Library of Science},
+  title = {Ten Simple Rules for Reproducible Computational Research},
+  year = {2013},
+  month = {10},
+  volume = {9},
+  url = {https://doi.org/10.1371/journal.pcbi.1003285},
+  pages = {1-4},
+  abstract = {},
+  number = {10},
+  doi = {10.1371/journal.pcbi.1003285}
+}
+
+@Manual{tibble,
+  title = {tibble: Simple Data Frames},
+  author = {Müller, Kirill AND Wickham, Hadley},
+  year = {2019},
+  note = {R package version 2.1.3},
+  url = {https://CRAN.R-project.org/package=tibble},
+}
+
+@Manual{stringr,
+  title = {stringr: Simple, Consistent Wrappers for Common String Operations},
+  author = {Wickham, Hadley},
+  year = {2019},
+  note = {R package version 1.4.0},
+  url = {https://cran.r-project.org/package=stringr},
+}
+
+@article{Sullivan:2019,
+  author = {Sullivan, Ian AND DeHaven, Alexander AND Mellor, David},
+  title = {Open and Reproducible Research on Open Science Framework},
+  journal = {Current protocols},
+  year = {2019},
+  volume = {18},
+  number = {1},
+  pages = {e32},
+  month = jun,
+  doi = {10.1002/cpet.32}
+}
+
+@article{Wilson:2017,
+  author = {Wilson, Greg AND Bryan, Jennifer AND Cranston, Karen AND Kitzes, Justin AND Nederbragt, Lex AND Teal, Tracy K.},
+  journal = {PLOS Computational Biology},
+  publisher = {Public Library of Science},
+  title = {Good enough practices in scientific computing},
+  year = {2017},
+  month = {06},
+  volume = {13},
+  url = {https://doi.org/10.1371/journal.pcbi.1005510},
+  pages = {1-20},
+  number = {6},
+  doi = {10.1371/journal.pcbi.1005510}
+}
