### 2020-01-15 #279 @aaronwolen
Thanks, @maelle! I added in the missing bib file and I'm working on updating the docs. Thanks for the Rmd fragment tip!
