### 2020-01-16 #279 @maelle
@HeidiSeibold Thanks, no problem, I'll re-read your review after @aaronwolen updates the docs.

@aaronwolen I'm using Rmd fragments in codemetar dev branch and really liking them. Feel free to comment on their use at https://github.com/ropensci/dev_guide/issues/159
