### 2020-01-19 a8ea871
Add info and resources to ?osfr-package
