### 2020-01-19 615f39d
Overhaul getting started vignette
- now centers around the classic gender/politics example
- utilizes multfile transfers
- provides links to useful OSF guides
- add a toc
