+++ vignettes/getting_started.Rmd
@@ -1,7 +1,9 @@
 ---
 title: "Getting Started with osfr"
 date: "`r Sys.Date()`"
-output: rmarkdown::html_vignette
+output: 
+  rmarkdown::html_vignette:
+    toc: TRUE
 vignette: >
   %\VignetteIndexEntry{Getting Started with osfr}
   %\VignetteEngine{knitr::rmarkdown}
@@ -13,121 +15,214 @@ knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
+
+Sys.unsetenv("OSF_SERVER")
 ```
 
-This vignette provides a quick tour of the `osfr` package. 
+This vignette provides a quick tour of the *osfr* package. 
 
 ```{r message=FALSE}
 library(osfr)
-library(dplyr)
-
-# external file for our example project
-write.csv(mtcars, "mtcars.csv")
 ```
 
+## What is OSF? 
 
-## Background 
+[OSF][osf] is a free and open source web application that provides a space for researchers to collaboratively store, manage, and share their research materials (e.g. data, code, protocols). 
 
-[OSF][osf], developed by the [Center for Open Science (COS)][cos], is a free, open source, web application that provides a space for researchers to collaboratively store, manage, and share their research materials (e.g. data, code, protocols). Files can be kept private, shared with a specific group of researchers, or made publicly available with citable DOIs so researchers can get credit for their work.
+Most work on OSF is organized around ***projects***, which include a cloud-based storage bucket where files can be stored and organized into directories. Note there is no storage limit on the size of projects but individual files must be < 5Gb. Projects can be kept private, shared with a specific group of collaborators, or made publicly available with citable DOIs so you can get credit for their work.
 
-`osfr` allows researchers to create and modify projects, components, and files through OSF's API.
+If you'd like to learn more about OSF the Center for Open Science has published an excellent series of [guides](http://help.osf.io/) to help get you started. We'll provide links to specific guides throughout this vignette.
+Here are a few relevant topics:
 
-## Working with projects/components
+* [Creating and Managing Projects][osf-projects]
+* [Adding sub-projects (i.e., components) to a project][osf-components]
+* [Understanding project permissions][osf-permissions]
 
-Most work on OSF is organized around ***projects***. Projects can contain *files*, groups of files in *directories*, and/or files in subsections of a project called ***components***. Collectively projects and components are called ***nodes***.
+## Accessing OSF projects
 
-### Creating a new project/component
+Let's check out an example project containing materials for an analysis of the 2012 American National Election Survey (ANES). You can access the OSF project in your browser by navigating to its URL: <https://osf.io/jgyxm/>. 
 
-At the start of a new research project or study we'll want to set up a new private project on OSF to store all the files related to the project. We can add multiple nested components or directories to create whatever structure will suite our workflow. Different components can have different DOIs and different privacy settings, so we may want to put similar files in their own components to organize the project.
-
-We'll first create a new *project* and store osfr's response in the `car_project` object. This object contains a specialized data frame based on the [tibble][] class that provides the new project's global unique identifier (GUID), title, and other metadata returned by the API. 
-
-*Note: You'll need to be authenticated to create `nodes`. See the `osf_auth()` function documentation or the `auth` vignette for more information.*
+Let's load this project into R with `osfr::osf_retrieve_node()`:
 
 ```{r}
-car_project <- osf_create_project("Motor Trend Car Road Tests")
-car_project
+anes_project <- osf_retrieve_node("https://osf.io/jgyxm")
+anes_project
 ```
 
[…]
