### 2020-01-20 4f5d531
Fix recursively deleting nodes
