### 2020-01-20 08a0e4a
bump dev version and update docs
