### 2020-01-20 7a83180
bump dev version and update docs
