### 2020-01-20 #279 @aaronwolen
Hi @maelle, I think I've addressed the remaining issues. Here's a quick summary of recent changes:

* added in this missing bibliography
* merged your PR to fix those 2 typos in the paper (Thanks!)
* vignette has been overhauled and now utilizes the multi-file upload/download features that were added after the original vignette was written
* vignette now includes information about projects/components, storage limits, and permissions, as well as links to relevant OSF guides
* added some information about OSF, projects/components and storage limits to `?osfr`, as well as links to the vignette and other helpful OSF guides

These are currently on the `dev` branch in case any additional changes are required, that way I can cut a single release upon merging to master.
