### 2020-01-20 7e32528
Update recurse argument in tests
