+++ tests/testthat/test-directories.R
@@ -10,7 +10,7 @@ setup({
 
 teardown({
   if (has_pat()) {
-    osf_rm(p1, recursive = TRUE, check = FALSE)
+    osf_rm(p1, recurse = TRUE, check = FALSE)
   }
 })
 
