### 2020-01-20 e221164
Update recurse argument in tests
