+++ tests/testthat/test-html-entity-encoding.R
@@ -12,7 +12,7 @@ setup({
 
 teardown({
   if (has_pat()) {
-    osf_rm(p1, recursive = TRUE, check = FALSE)
+    osf_rm(p1, recurse = TRUE, check = FALSE)
   }
 })
 
