### 2020-01-20 089aa00
Tweak two vignette section titles
