### 2020-01-20 d7c9c39
Use osf_dev_off() when building the vignette
