### 2020-01-20 420896c
Improve description of osf_tbls in the paper
