### 2020-01-21 #279 @maelle
Approved, the changes look good! Thanks @aaronwolen  for submitting and @HeidiSeibold @cboettig  for your reviews! :smiley_cat:

- [ ] (Suggestion) In http://centerforopenscience.github.io/osfr/articles/auth.html#using-your-pat you could add more links, e.g. to https://rstats.wtf/r-startup.html, https://usethis.r-lib.org/reference/edit.html.
- [ ] Merge dev into master.
"```
- [ ] If you already had a `pkgdown` website, fix its URL to point to `https://docs.ropensci.org/package_name` and deactivate the automatic deployment you might have set up, since it will not be built centrally like for all rOpenSci packages, see http://devdevguide.netlify.com/#docsropensci (it might take 1-2 days for the system to pick it up). In addition, in your DESCRIPTION file, include the docs link in the `URL` field alongside the link to the GitHub repository, e.g.: `URL: https://docs.ropensci.org/foobar (website) https://github.com/ropensci/foobar` **I think you might need to read [Jeroen's note about precomputing vignettes](https://ropensci.org/technotes/2019/12/08/precompute-vignettes/) since the docs server can't access credentials.**
- [ ] Add a mention of the review in `DESCRIPTION` via [`rodev::add_ro_desc()`](https://docs.ropensci.org/rodev/reference/add_ro_desc.html).
- [ ] Submit to JOSS at <https://joss.theoj.org/papers/new>, using the rOpenSci GitHub repo URL. When a JOSS "PRE REVIEW" issue is generated for your paper, add the comment: `This package has been reviewed by rOpenSci: https://LINK.TO/THE/REVIEW/ISSUE, @ropensci/editors`

Welcome aboard! We'd love to host a blog post about your package - either a [short introduction to it with one example](https://ropensci.org/tech-notes/) or [a longer post with some narrative about its development or something you learned, and an example of its use](https://ropensci.org/blog/). If you are interested, review [the instructions](https://github.com/ropensci/roweb2#contributing-a-blog-post), and tag `@stefaniebutland` in your reply. She will get in touch about timing and can answer any questions.
