### 2020-01-21 #279 @aaronwolen
@maelle, COS has okayed transferring the repo but it will have to be handled by the admin for their organization. Could you add @mfraezz to the ropensci team so he can make the move?
