### 2020-01-21 #279 @maelle
Thanks, done!

Once the repo is transferred so that I might give both of you admin rights again (note that my timezone is UTC+1).
