### 2020-01-21 #279 @mfraezz
I'm an owner of the CenterForOpenScience organization, so it was implicit before. I'd recommend updating the repo description to include the new GH pages url, https://ropensci.github.io/osfr/ , but otherwise have no real need for access.
