### 2020-01-21 1127a19
auth vignette: link to rstats.wtf
