### 2020-01-21 3915752
Add ropensci footer to readme
