### 2020-01-21 #279 @aaronwolen
@cboettig and @HeidiSeibold, any objections to being listed as reviewers in the package `DESCRIPTION`?
