### 2020-01-21 ee253b7
Update links to point to ropensci repo and docs
