+++ R/utils-api-responses.R
@@ -23,7 +23,7 @@ process_response <- function(res) {
   if (res$status_code > 500) {
     abort(paste0(
       "Encountered an unexpected error with the OSF API\n",
-      "Please report this at https://github.com/centerforopenscience/osfr/issues\n",
+      "Please report this at https://github.com/ropensci/osfr/issues\n",
       "* Status code: ", res$status_code, "\n",
       "* Request: ", res$request$url$url, "\n",
       "* Resonse: \n", res$parse("UTF-8")
