### 2020-01-21 52015b4
Update links to point to ropensci repo and docs
