### 2020-01-21 583b042
Remove pkgdown bootswatch template
