### 2020-01-21 d9ac8e5
precompute the getting started vignette
suggested by @maelle and documented here:
https://ropensci.org/technotes/2019/12/08/precompute-vignettes/
