### 2020-01-21 47d24d8
Update docs and bump version to 0.2.7
