+++ man/osfr-package.Rd
@@ -38,9 +38,9 @@ the Center for Open Science: \url{https://help.osf.io}
 \seealso{
 Useful links:
 \itemize{
-  \item \url{http://centerforopenscience.github.io/osfr}
-  \item \url{https://github.com/centerforopenscience/osfr}
-  \item Report bugs at \url{https://github.com/centerforopenscience/osfr/issues}
+  \item \url{https://docs.ropensci.org/osfr}
+  \item \url{https://github.com/ropensci/osfr}
+  \item Report bugs at \url{https://github.com/ropensci/osfr/issues}
 }
 
 }
