+++ README.md
@@ -1,21 +1,20 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-# osfr <a href="http://centerforopenscience.github.io/osfr"><img src="man/figures/logo.png" align="right" height="139" /></a>
+# osfr <a href="http://ropensci/osfr.io/osfr"><img src="man/figures/logo.png" align="right" height="139" /></a>
 
 [![Build
-Status](https://travis-ci.org/CenterForOpenScience/osfr.svg?branch=master)](https://travis-ci.org/CenterForOpenScience/osfr)
+Status](https://travis-ci.org/ropensci/osfr.svg?branch=master)](https://travis-ci.org/ropensci/osfr)
 [![AppVeyor build
-status](https://ci.appveyor.com/api/projects/status/github/CenterForOpenScience/osfr?branch=master&svg=true)](https://ci.appveyor.com/project/CenterForOpenScience24999/osfr)
+status](https://ci.appveyor.com/api/projects/status/github/ropensci/osfr?branch=master&svg=true)](https://ci.appveyor.com/project/aaronwolen/osfr)
 [![Coverage
-status](https://codecov.io/gh/CenterForOpenScience/osfr/branch/master/graph/badge.svg)](https://codecov.io/github/CenterForOpenScience/osfr?branch=master)
+status](https://codecov.io/gh/ropensci/osfr/branch/master/graph/badge.svg)](https://codecov.io/github/ropensci/osfr?branch=master)
 [![](https://badges.ropensci.org/279_status.svg)](https://github.com/ropensci/software-review/issues/279)
 
 ## Overview
 
-osfr provides a suite of functions for interacting with
-[OSF](https://osf.io "Open Science Framework") (<https://osf.io>) that
-are primarily focused on project management workflows.
+osfr provides a suite of functions for interacting with the Open Science
+Framework ([OSF](https://osf.io "Open Science Framework")).
 
 **What is OSF?**
 
@@ -34,14 +33,14 @@ CRAN. You can install the current version with the *remotes* package:
 
 ``` r
 # install.packages("remotes")
-remotes::install_github("centerforopenscience/osfr")
+remotes::install_github("ropensci/osfr")
 ```
 
 ## Usage Examples
 
 *Note: You need to [setup an OSF personal access token
-(PAT)](http://centerforopenscience.github.io/osfr/articles/auth.html) to
-use osfr to manage projects or upload files.*
+(PAT)](http://ropensci/osfr.io/osfr/articles/auth.html) to use osfr to
+manage projects or upload files.*
 
 ### Accessing Open Research Materials
 
@@ -64,9 +63,9 @@ library(osfr)
 cr_project <- osf_retrieve_node("e81xl")
 cr_project
 #> # A tibble: 1 x 3
-#>   name                                    id    meta      
-#>   <chr>                                   <chr> <list>    
-#> 1 Reproducibility Project: Cancer Biology e81xl <list [3]>
+#>   name                                    id    meta            
+#>   <chr>                                   <chr> <list>          
+#> 1 Reproducibility Project: Cancer Biology e81xl <named list [3]>
 ```
 
 This returns an `osf_tbl` object with a single row representing the
@@ -76,12 +75,12 @@ project.
 ``` r
 osf_ls_files(cr_project)
 #> # A tibble: 4 x 3
-#>   name                                     id                     meta     
-#>   <chr>                                    <chr>                  <list>   
-#> 1 Adjustment of 50 studies to 37 studies.… 565602398c5e4a3877d72… <list [3…
-#> 2 papers_and_keywords.xlsx                 553e671b8c5e4a219919e… <list [3…
-#> 3 Full_dataset_of_papers_formatted.xls     553e671b8c5e4a219919e… <list [3…
[…]
