### 2020-01-21 926cc2d
Update docs and bump version to 0.2.7
