### 2020-01-21 34f743f
Add script to precompute vignettes
