+++ vignettes/precompile.R
@@ -1,0 +1,3 @@
+# Precompiled vignettes that depend on API key
+library(knitr)
+knit("vignettes/getting_started.Rmd.orig", "vignettes/getting_started.Rmd")
