### 2020-01-22 #279 @HeidiSeibold
> @cboettig and @HeidiSeibold, any objections to being listed as reviewers in the package DESCRIPTION?

I'd be honored :)
