+++ tests/testthat/test-directories.R
@@ -1,50 +1,57 @@
 context("Directories")
 
 # setup -------------------------------------------------------------------
-setup({
-  vcr::insert_cassette("test-directories", record = "new_episodes")
-  if (has_pat()) {
-    p1 <<- osf_create_project(title = "osfr-test-directories")
-  }
-})
+vcr_config("cassettes/directories")
+# record <- "new_episodes"
+# record <- "none"
+record <- "once"
 
-teardown({
+setup({
   if (has_pat()) {
-    osf_rm(p1, recurse = TRUE, check = FALSE)
+    vcr::use_cassette("create-p5", record = record, {
+      p1 <<- osf_create_project(title = "osfr-test-directories3")
+    })
   }
-  vcr::eject_cassette("test-directories")
 })
 
 
 # tests -------------------------------------------------------------------
 test_that("empty project/folder returns a osf_tbl_file with 0 rows", {
   skip_if_no_pat()
+  vcr::use_cassette("list-empty-p1", record = record, {
+    out <- osf_ls_files(p1)
+  })
 
-  out <- osf_ls_files(p1)
   expect_s3_class(out, "osf_tbl_file")
   expect_equal(nrow(out), 0)
 })
 
 test_that("listing a non-existent folder errors", {
-  vcr::insert_cassette("test-directories-nonexistent"), record = "new_episodes")
   skip_if_no_pat()
-  expect_error(osf_ls_files(p1, path = "nonexistent"), "Can't find directory")
-  vcr::eject_cassette()
+  expect_error(
+    vcr::use_cassette("list-nonexistent-dir", record = record, {
+      osf_ls_files(p1, path = "nonexistent")
+    }),
+    "Can't find directory"
+  )
 })
 
 test_that("create a top-level directory", {
   skip_if_no_pat()
+  vcr::use_cassette("create-dir1", record = record, {
+    d1 <- osf_mkdir(p1, path = "dir1")
+  })
 
-  d1 <- osf_mkdir(p1, path = "dir1")
   expect_s3_class(d1, "osf_tbl_file")
   expect_equal(d1$name, "dir1")
 })
 
 test_that("list a top-level directory", {
   skip_if_no_pat()
+  vcr::use_cassette("list-p1-with-one-dir", record = record, {
+    out <- osf_ls_files(p1)
+  })
 
-  p1_nonempty <- p1
-  out <- osf_ls_files(p1_nonempty)
   expect_s3_class(out, "osf_tbl_file")
   expect_equal(nrow(out), 1)
   expect_equal(out$name, "dir1")
@@ -53,29 +60,30 @@ test_that("list a top-level directory", {
 
 test_that("create a subdirectory within an existing directory", {
   skip_if_no_pat()
+  vcr::use_cassette("create-subdir-dir11", record = record, {
+    d11 <- osf_mkdir(p1, path = "dir1/dir11")
+  })
 
-  d11 <- osf_mkdir(p1, path = "dir1/dir11")
   expect_s3_class(d11, "osf_tbl_file")
   expect_equal(d11$name, "dir11")
 })
 
-test_that("list a subdirectory", {
+test_that("list within a subdirectory", {
   skip_if_no_pat()
+  vcr::use_cassette("list-within-subdir-dir1", record = record, {
+    out <- osf_ls_files(p1, path = "dir1")
+  })
 
-  p1_withsubdir <- p1
-  out <- osf_ls_files(p1_withsubdir)
-  expect_equal(nrow(out), 1)
-  expect_equal(out$name, "dir1")
-
-  out <- osf_ls_files(p1_withsubdir, path = "dir1")
   expect_equal(nrow(out), 1)
[…]
