+++ tests/testthat/helper-vcr.R
@@ -1,9 +1,13 @@
 library(vcr)
 
-invisible(
+# redefine same global settings until the following issue is resolved
+# https://github.com/ropensci/vcr/issues/136
+vcr_config <- function(dir, record = "new_episodes") {
+  invisible(
   vcr::vcr_configure(
-    dir = "tests/testthat/cassettes",
-    match_requests_on = c("method", "uri", "body"),
+    dir = file.path(rprojroot::find_testthat_root_file(), dir),
+    record = record,
+    match_requests_on = c("method", "uri"),
     filter_sensitive_data = list("<totallyrealpat>" = Sys.getenv("OSF_PAT")),
     log = nzchar(Sys.getenv("VCR_LOG")),
     log_opts = list(
@@ -13,4 +17,4 @@ invisible(
     )
   )
 )
-
+}
