### 2020-01-22 0c4c52d
add missing skip_if_no_pat()
