### 2020-01-22 25fc0a7
Remove travis deploy step
Website is now build by ropensci
