+++ tests/testthat/cassettes/directories/create-dir.yml
@@ -1,0 +1,112 @@
+http_interactions:
+- request:
+    method: get
+    uri: https://api.test.osf.io/v2/nodes/gyrp6/files/osfstorage/?filter%5Bkind%5D=folder&filter%5Bname%5D=dir&page=1
+    body:
+      encoding: ''
+      string: ''
+    headers:
+      Accept-Encoding: gzip, deflate
+      Accept: application/json, text/xml, application/xml, */*
+      User-Agent: osfr v0.2.7.9000
+      Authorization: Bearer <totallyrealpat>
+      Accept-Header: application/vnd.api+json;version=2.8
+  response:
+    status:
+      status_code: '200'
+      message: OK
+      explanation: Request fulfilled, document follows
+    headers:
+      allow: GET, HEAD, OPTIONS
+      alt-svc: clear
+      cache-control: no-cache, no-store, max-age=0, must-revalidate
+      content-encoding: gzip
+      content-type: application/vnd.api+json; charset=utf-8
+      date: Wed, 22 Jan 2020 20:56:08 GMT
+      expires: Mon, 01 Jan 1990 00:00:00 GMT
+      pragma: no-cache
+      server: nginx
+      status: HTTP/1.1 200 OK
+      transfer-encoding: chunked
+      vary:
+      - Accept-Encoding
+      - Accept
+      via: 1.1 google
+      x-frame-options: SAMEORIGIN
+    body:
+      encoding: UTF-8
+      file: no
+      string: '{"data":[{"id":"5e28b6cf68e7e20009f477c0","type":"files","attributes":{"guid":null,"checkout":null,"name":"dir1","kind":"folder","path":"/5e28b6cf68e7e20009f477c0/","size":null,"provider":"osfstorage","materialized_path":"/dir1/","last_touched":null,"date_modified":null,"date_created":null,"extra":{"hashes":{"md5":null,"sha256":null}},"tags":[],"current_user_can_comment":true,"current_version":1},"relationships":{"parent_folder":{"links":{"related":{"href":"https://api.test.osf.io/v2/files/5e28b6907f42440009faeb74/","meta":{}}},"data":{"id":"5e28b6907f42440009faeb74","type":"files"}},"files":{"links":{"related":{"href":"https://api.test.osf.io/v2/nodes/gyrp6/files/osfstorage/5e28b6cf68e7e20009f477c0/","meta":{}}}},"node":{"links":{"related":{"href":"https://api.test.osf.io/v2/nodes/gyrp6/","meta":{}}},"data":{"id":"gyrp6","type":"nodes"}},"target":{"links":{"related":{"href":"https://api.test.osf.io/v2/nodes/gyrp6/","meta":{"type":"node"}}},"data":{"type":"node","id":"gyrp6"}}},"links":{"info":"https://api.test.osf.io/v2/files/5e28b6cf68e7e20009f477c0/","move":"https://files.us.test.osf.io/v1/resources/gyrp6/providers/osfstorage/5e28b6cf68e7e20009f477c0/","upload":"https://files.us.test.osf.io/v1/resources/gyrp6/providers/osfstorage/5e28b6cf68e7e20009f477c0/","delete":"https://files.us.test.osf.io/v1/resources/gyrp6/providers/osfstorage/5e28b6cf68e7e20009f477c0/","new_folder":"https://files.us.test.osf.io/v1/resources/gyrp6/providers/osfstorage/5e28b6cf68e7e20009f477c0/?kind=folder","self":"https://api.test.osf.io/v2/files/5e28b6cf68e7e20009f477c0/"}},{"id":"5e28b6e09ee3270009ec10d9","type":"files","attributes":{"guid":null,"checkout":null,"name
[…]
