### 2020-01-22 7cae6fa
Increase travis timeout to 30min (for shame)
