+++ NEWS.md
@@ -2,7 +2,7 @@
 
 ## Important changes
 
-osfr is now part of rOpenSci and the documentation website has moved to a new URL:  <https://docs.ropensci.org/osfr>.
+osfr is now part of rOpenSci and the documentation website has moved to a new URL:  <https://ropensci.github.io/osfr>.
 
 ## New features
 
@@ -115,5 +115,5 @@ previous package can be installed with the *remotes* package:
 remotes::install_github("ropensci/osfr@v0.1.1")
 ```
 
-See <https://docs.ropensci.org/osfr/> for details about the new
+See <https://ropensci.github.io/osfr> for details about the new
 package.
