### 2020-01-23 6bcdc82
Downgrade logger warning to startup message
Also move from .onLoad() to .onAttach() to avoid R CMD CHECK warning:
* checking whether package ‘osfr’ can be installed ... WARNING
Found the following significant warnings:
  Warning: The logger package must be installed to enable logging
