### 2020-01-23 e1aa15a
Remove deleted project's URL from vignette
Fixes R CMD CHECK warning
