### 2020-01-23 fc04610
Add cran-comments
