+++ cran-comments.md
@@ -1,0 +1,17 @@
+## Test environments
+* local OS X install, R 3.6
+* ubuntu 14.04 (on travis-ci),  R 3.5.1, devel and release
+* win-builder (devel and release)
+* Windows Server 2008 R2 SP1 (on R-hub), R-devel
+* Ubuntu Linux 16.04 LTS + GCD (on R-hub), R-release
+* Fedora Linux + clang + gfortran (on R-hub), R-devel
+
+## R CMD check results
+
+0 errors | 0 warnings | 0 notes
+
+* This is a new release.
+
+## Reverse dependencies
+
+There are no reverse dependencies.
