### 2020-01-23 21a21cc
Bump version to 0.2.8
