+++ man-roxygen/synchronization.R
@@ -5,5 +5,5 @@
 #' something like [`rsync`](https://rsync.samba.org),
 #' `osf_download()`/`osf_upload()` do not take into account a file's contents or
 #' modification time. Whether you're uploading or downloading, if `overwrite =
-#' TRUE`, osfr will overwrite an existing file regradless of whether the
+#' TRUE`, osfr will overwrite an existing file regardless of whether the
 #' existing file is the more recent copy. You have been warned.
