### 2020-01-24 14728cd
Fix coc and contrib links in readme
