### 2020-01-24 e0f5009
Update cran-comments
