+++ cran-comments.md
@@ -10,7 +10,11 @@
 
 0 errors | 0 warnings | 0 notes
 
-* This is a new release.
+## Resubmission
+
+This is a resubmission. In this version I have:
+
+- replaced links in README.md with fully specified URLs
 
 ## Reverse dependencies
 
