### 2020-01-28 078af63
Restore URLs to new ropensci osfr website
