+++ DESCRIPTION
@@ -32,6 +32,7 @@ Imports:
     httr
 License: Apache License 2.0 | file LICENSE
 URL:
+    https://docs.ropensci.org/osfr,
     https://github.com/ropensci/osfr
 BugReports: https://github.com/ropensci/osfr/issues
 LazyData: true
