+++ tests/testthat/cassettes/directories/create-subdir-dir21.yml
@@ -1,14 +1,14 @@
 http_interactions:
 - request:
     method: get
-    uri: https://api.test.osf.io/v2/nodes/gyrp6/files/osfstorage/?filter%5Bkind%5D=folder&filter%5Bname%5D=dir2&page=1
+    uri: https://api.test.osf.io/v2/nodes/mdfb8/files/osfstorage/?filter%5Bkind%5D=folder&filter%5Bname%5D=dir2&page=1
     body:
       encoding: ''
       string: ''
     headers:
       Accept-Encoding: gzip, deflate
       Accept: application/json, text/xml, application/xml, */*
-      User-Agent: osfr v0.2.7.9000
+      User-Agent: osfr v0.2.8
       Authorization: Bearer <totallyrealpat>
       Accept-Header: application/vnd.api+json;version=2.8
   response:
@@ -22,7 +22,7 @@ http_interactions:
       cache-control: no-cache, no-store, max-age=0, must-revalidate
       content-length: '128'
       content-type: application/vnd.api+json; charset=utf-8
-      date: Wed, 22 Jan 2020 20:55:59 GMT
+      date: Wed, 29 Jan 2020 23:27:49 GMT
       expires: Mon, 01 Jan 1990 00:00:00 GMT
       pragma: no-cache
       server: nginx
@@ -36,49 +36,60 @@ http_interactions:
       encoding: UTF-8
       file: no
       string: '{"data":[],"links":{"first":null,"last":null,"prev":null,"next":null,"meta":{"total":0,"per_page":10}},"meta":{"version":"2.0"}}'
-  recorded_at: 2020-01-22 20:56:06 GMT
-  recorded_with: vcr/0.4.1.93, webmockr/0.5.1.96
+  recorded_at: 2020-01-29 23:27:53 GMT
+  recorded_with: vcr/0.4.1.93.9000, webmockr/0.5.1.96
 - request:
     method: put
-    uri: https://files.us.test.osf.io/v1/resources/gyrp6/providers/osfstorage/?kind=folder&name=dir2
+    uri: https://files.us.test.osf.io/v1/resources/mdfb8/providers/osfstorage/?kind=folder&name=dir2
     body:
       encoding: ''
       string: ''
     headers:
       Accept-Encoding: gzip, deflate
       Accept: application/json, text/xml, application/xml, */*
-      User-Agent: osfr v0.2.7.9000
+      User-Agent: osfr v0.2.8
       Authorization: Bearer <totallyrealpat>
   response:
     status:
-      status_code: '409'
-      message: Conflict
-      explanation: Request conflict.
+      status_code: '201'
+      message: Created
+      explanation: Document created, URL follows
     headers:
-      status: HTTP/1.1 409 Conflict
+      status: HTTP/1.1 201 Created
       server: nginx
-      date: Wed, 22 Jan 2020 20:56:01 GMT
+      date: Wed, 29 Jan 2020 23:27:49 GMT
       content-type: application/json; charset=UTF-8
-      content-length: '107'
+      content-length: '821'
+      x-waterbutler-request-id: 8fe7db0f-0559-40ad-93ae-c228ae956318
+      cache-control: no-cache, no-store, max-age=0, must-revalidate
+      expires: Mon, 01 Jan 1990 00:00:00 GMT
+      pragma: no-cache
       via: 1.1 google
       alt-svc: clear
     body:
       encoding: UTF-8
       file: no
-      string: '{"message": "Cannot complete action: file or folder \"dir2\" already
[…]
