### 2020-01-30 65ef1f8
Relocate test-directories teardown block
