### 2020-01-30 c782895
use vcr for multi-file upload tests
