+++ tests/testthat/cassettes/uploading-multiple-files/list-c4-subdir1_1-files.yml
@@ -1,0 +1,121 @@
+http_interactions:
+- request:
+    method: get
+    uri: https://api.test.osf.io/v2/nodes/xh6v8/files/osfstorage/5e33479404172e0009b161c9/?filter%5Bkind%5D=folder&filter%5Bname%5D=subdir1&page=1
+    body:
+      encoding: ''
+      string: ''
+    headers:
+      Accept-Encoding: gzip, deflate
+      Accept: application/json, text/xml, application/xml, */*
+      User-Agent: osfr v0.2.8
+      Authorization: Bearer <totallyrealpat>
+      Accept-Header: application/vnd.api+json;version=2.8
+  response:
+    status:
+      status_code: '200'
+      message: OK
+      explanation: Request fulfilled, document follows
+    headers:
+      allow: GET, HEAD, OPTIONS
+      alt-svc: clear
+      cache-control: no-cache, no-store, max-age=0, must-revalidate
+      content-encoding: gzip
+      content-type: application/vnd.api+json; charset=utf-8
+      date: Thu, 30 Jan 2020 21:17:47 GMT
+      expires: Mon, 01 Jan 1990 00:00:00 GMT
+      pragma: no-cache
+      server: nginx
+      status: HTTP/1.1 200 OK
+      transfer-encoding: chunked
+      vary:
+      - Accept-Encoding
+      - Accept
+      via: 1.1 google
+      x-frame-options: SAMEORIGIN
+    body:
+      encoding: UTF-8
+      file: no
+      string: '{"data":[{"id":"5e334796342fc90009b70d1c","type":"files","attributes":{"guid":null,"checkout":null,"name":"subdir1","kind":"folder","path":"/5e334796342fc90009b70d1c/","size":null,"provider":"osfstorage","materialized_path":"/.osfr-tests/subdir1/","last_touched":null,"date_modified":null,"date_created":null,"extra":{"hashes":{"md5":null,"sha256":null}},"tags":[],"current_user_can_comment":true,"current_version":1},"relationships":{"parent_folder":{"links":{"related":{"href":"https://api.test.osf.io/v2/files/5e33479404172e0009b161c9/","meta":{}}},"data":{"id":"5e33479404172e0009b161c9","type":"files"}},"files":{"links":{"related":{"href":"https://api.test.osf.io/v2/nodes/xh6v8/files/osfstorage/5e334796342fc90009b70d1c/","meta":{}}}},"node":{"links":{"related":{"href":"https://api.test.osf.io/v2/nodes/xh6v8/","meta":{}}},"data":{"id":"xh6v8","type":"nodes"}},"target":{"links":{"related":{"href":"https://api.test.osf.io/v2/nodes/xh6v8/","meta":{"type":"node"}}},"data":{"type":"node","id":"xh6v8"}}},"links":{"info":"https://api.test.osf.io/v2/files/5e334796342fc90009b70d1c/","move":"https://files.us.test.osf.io/v1/resources/xh6v8/providers/osfstorage/5e334796342fc90009b70d1c/","upload":"https://files.us.test.osf.io/v1/resources/xh6v8/providers/osfstorage/5e334796342fc90009b70d1c/","delete":"https://files.us.test.osf.io/v1/resources/xh6v8/providers/osfstorage/5e334796342fc90009b70d1c/","new_folder":"https://files.us.test.osf.io/v1/resources/xh6v8/providers/osfstorage/5e334796342fc90009b70d1c/?kind=folder","self":"https://api.test.osf.io/v2/files/5e334796342fc90009b70d1c/"}}],"links":{"first":null,"last":null,"prev":null,"next":null,"meta":
[…]

[…]
