### 2020-01-30 8e044b2
use vcr for multi-file upload tests
