+++ tests/testthat/test-uploading-multiple-files.R
@@ -2,6 +2,7 @@ context("Uploading multiple files")
 
 
 # setup -------------------------------------------------------------------
+vcr_config("cassettes/uploading-multiple-files")
 
 setup({
   # copy directory for testing multifile uploads to pwd
@@ -11,14 +12,18 @@ setup({
   )
 
   if (has_pat()) {
-    p1 <<- osf_create_project(title = "osfr-multifile-upload-test")
+    vcr::use_cassette("create-p1", record = record, {
+      p1 <<- osf_create_project(title = "osfr-multifile-upload-test")
+    })
   }
 })
 
 teardown({
   fs::dir_delete(multidir)
   if (has_pat()) {
-    osf_rm(p1, recurse = TRUE, check = FALSE)
+    vcr::use_cassette("delete-p1", record = record, {
+      osf_rm(p1, recurse = TRUE, check = FALSE)
+    })
   }
 })
 
@@ -29,7 +34,11 @@ test_that("multiple files can be uploaded", {
   skip_if_no_pat()
 
   infiles <- fs::dir_ls(multidir, type = "file")
-  out <- osf_upload(p1, infiles)
+
+  vcr::use_cassette("upload-infiles", record = record, {
+    out <- osf_upload(p1, infiles)
+  })
+
   expect_s3_class(out, "osf_tbl_file")
   expect_equal(out$name, basename(infiles))
 })
@@ -37,71 +46,82 @@ test_that("multiple files can be uploaded", {
 test_that("a directory can be uploaded", {
   skip_if_no_pat()
 
-  c1 <- osf_create_component(p1, "dir-upload")
-  out <- osf_upload(c1, multidir)
+  vcr::use_cassette("create-and-populate-c1", record = record, {
+    c1 <- osf_create_component(p1, "dir-upload")
+    out <- osf_upload(c1, multidir)
+  })
   expect_s3_class(out, "osf_tbl_file")
   expect_match(out$name, basename(multidir))
 
   # files within the directory were uploaded
-  expect_equal(
-    osf_ls_files(out, type = "file")$name,
-    basename(fs::dir_ls(multidir, type = "file"))
+  vcr::use_cassette("list-c1-files", record = record,
+    c1_files <- osf_ls_files(out)
   )
-
-  # subdirectories within the directory were created
-  subdirs <- basename(fs::dir_ls(multidir, type = "directory"))
-  expect_equal(
-    osf_ls_files(out, type = "folder")$name,
-    subdirs
+  expect_true(
+    all(c1_files$name %in% basename(fs::dir_ls(multidir)))
   )
-
-  # subdirectories are empty with recurse=FALSE
-  subdir1 <- osf_ls_files(out, path = subdirs[1])
-  expect_equal(nrow(subdir1), 0)
 })
 
 test_that("a subdirectory can be uploaded", {
   skip_if_no_pat()
 
   # upload the subdirectory with no children
-  c2 <- osf_create_component(p1, "subdir-upload")
-
   indir <- file.path(multidir, "subdir1", "subdir1_1")
-  out <- osf_upload(c2, indir, verbose = TRUE)
+  vcr::use_cassette("create-and-populate-c2", record = record, {
+    c2 <- osf_create_component(p1, "subdir-upload")
+    out <- osf_upload(c2, indir, verbose = TRUE)
+  })
 
   expect_s3_class(out, "osf_tbl_file")
   expect_match(out$name, basename(indir))
 
   # verify leaf directory was created in the project root
-  expect_match(
-    osf_ls_files(c2, type = "folder")$name,
+  vcr::use_cassette("list-c2-subdirs", record = record,
[…]
