+++ tests/testthat/cassettes/uploading-multiple-files/c5-upload-updir.yml
@@ -1,0 +1,207 @@
+http_interactions:
+- request:
+    method: get
+    uri: https://api.test.osf.io/v2/nodes/bdhyq/files/osfstorage/?filter%5Bkind%5D=folder&filter%5Bname%5D=subdir2&page=1
+    body:
+      encoding: ''
+      string: ''
+    headers:
+      Accept-Encoding: gzip, deflate
+      Accept: application/json, text/xml, application/xml, */*
+      User-Agent: osfr v0.2.8
+      Authorization: Bearer <totallyrealpat>
+      Accept-Header: application/vnd.api+json;version=2.8
+  response:
+    status:
+      status_code: '200'
+      message: OK
+      explanation: Request fulfilled, document follows
+    headers:
+      allow: GET, HEAD, OPTIONS
+      alt-svc: clear
+      cache-control: no-cache, no-store, max-age=0, must-revalidate
+      content-length: '128'
+      content-type: application/vnd.api+json; charset=utf-8
+      date: Thu, 30 Jan 2020 21:18:25 GMT
+      expires: Mon, 01 Jan 1990 00:00:00 GMT
+      pragma: no-cache
+      server: nginx
+      status: HTTP/1.1 200 OK
+      vary:
+      - Accept-Encoding
+      - Accept
+      via: 1.1 google
+      x-frame-options: SAMEORIGIN
+    body:
+      encoding: UTF-8
+      file: no
+      string: '{"data":[],"links":{"first":null,"last":null,"prev":null,"next":null,"meta":{"total":0,"per_page":10}},"meta":{"version":"2.0"}}'
+  recorded_at: 2020-01-30 21:18:29 GMT
+  recorded_with: vcr/0.4.1.93, webmockr/0.5.1.96
+- request:
+    method: put
+    uri: https://files.us.test.osf.io/v1/resources/bdhyq/providers/osfstorage/?kind=folder&name=subdir2
+    body:
+      encoding: ''
+      string: ''
+    headers:
+      Accept-Encoding: gzip, deflate
+      Accept: application/json, text/xml, application/xml, */*
+      User-Agent: osfr v0.2.8
+      Authorization: Bearer <totallyrealpat>
+  response:
+    status:
+      status_code: '201'
+      message: Created
+      explanation: Document created, URL follows
+    headers:
+      status: HTTP/1.1 201 Created
+      server: nginx
+      date: Thu, 30 Jan 2020 21:18:27 GMT
+      content-type: application/json; charset=UTF-8
+      content-length: '827'
+      x-waterbutler-request-id: 567eb9fd-ef0b-400f-96d7-d2f2295daa1f
+      cache-control: no-cache, no-store, max-age=0, must-revalidate
+      expires: Mon, 01 Jan 1990 00:00:00 GMT
+      pragma: no-cache
+      via: 1.1 google
+      alt-svc: clear
+    body:
+      encoding: UTF-8
+      file: no
+      string: '{"data": {"type": "files", "links": {"upload": "https://files.us.test.osf.io/v1/resources/bdhyq/providers/osfstorage/5e33482304172e0009b161ed/?kind=file",
+        "new_folder": "https://files.us.test.osf.io/v1/resources/bdhyq/providers/osfstorage/5e33482304172e0009b161ed/?kind=folder",
+        "delete": "https://files.us.test.osf.io/v1/resources/bdhyq/providers/osfstorage/5e33482304172e0009b161ed/",
+        "move": "https://files.us.test.osf.io/v1/resources/bdhyq/providers/osfstorage/5e33482304172e0009b161ed/"},
[…]

[…]
