### 2020-02-04 #279 @maelle
@aaronwolen now that you've submitted the package to JOSS I think the only missing item from my list is your adding a codemeta.json to the repo (via `codemetar::write_codemeta()`, I recommend the dev version `remotes::install_github("ropensci/codemetar", ref = "dev")`).
