### 2020-02-04 a57d6c9
Start cassettes for downloading tests
- add write_disk_path argument to vcr_config() helper
