+++ tests/testthat/helper-vcr.R
@@ -5,12 +5,19 @@ record <- "once"
 
 # redefine same global settings until the following issue is resolved
 # https://github.com/ropensci/vcr/issues/136
-vcr_config <- function(dir, record = "new_episodes") {
+vcr_config <- function(dir, record = "new_episodes", write_disk_path = NULL, ...) {
+  rootdir <- rprojroot::find_testthat_root_file()
+
+  if (!is.null(write_disk_path)) {
+    write_disk_path <- file.path(rootdir, write_disk_path)
+  }
+
   invisible(
   vcr::vcr_configure(
-    dir = file.path(rprojroot::find_testthat_root_file(), dir),
+    dir = file.path(rootdir, dir),
     record = record,
     match_requests_on = c("method", "uri"),
+    write_disk_path = write_disk_path,
     filter_sensitive_data = list("<totallyrealpat>" = Sys.getenv("OSF_PAT")),
     log = nzchar(Sys.getenv("VCR_LOG")),
     log_opts = list(
@@ -19,5 +26,4 @@ vcr_config <- function(dir, record = "new_episodes") {
       date = TRUE
     )
   )
-)
-}
+)}
