### 2020-02-04 #279 @aaronwolen
Thanks, @maelle! Just added the missing `codemeta.json` to master. Meanwhile, I'm waiting to hear back from CRAN 🤞.
