### 2020-02-04 d03f5bc
Fix capitalization issue in paper.bib [skip ci]
