+++ inst/paper.bib
@@ -8,14 +8,14 @@
 
 @Manual{googledrive,
   title = {googledrive: An Interface to Google Drive},
-  author = {{D'Agostino McGowan}, Lucy AND Bryan, Jennifer},
+  author = {{D'Agostino McGowan}, Lucy and Bryan, Jennifer},
   year = {2019},
   note = {R package version 1.0.0},
   url = {https://cran.r-project.org/package=googledrive},
 }
 
 @Manual{magrittr,
-    title = {magrittr: A Forward-Pipe Operator for R},
+    title = {magrittr: A Forward-Pipe Operator for {R}},
     author = {Stefan Milton Bache and Hadley Wickham},
     year = {2014},
     note = {R package version 1.5},
@@ -34,7 +34,7 @@
 }
 
 @article{Sandve:2013,
-  author = {Sandve, Geir Kjetil AND Nekrutenko, Anton AND Taylor, James AND Hovig, Eivind},
+  author = {Sandve, Geir Kjetil and Nekrutenko, Anton and Taylor, James and Hovig, Eivind},
   journal = {PLOS Computational Biology},
   publisher = {Public Library of Science},
   title = {Ten Simple Rules for Reproducible Computational Research},
@@ -50,7 +50,7 @@
 
 @Manual{tibble,
   title = {tibble: Simple Data Frames},
-  author = {Müller, Kirill AND Wickham, Hadley},
+  author = {Müller, Kirill and Wickham, Hadley},
   year = {2019},
   note = {R package version 2.1.3},
   url = {https://CRAN.R-project.org/package=tibble},
@@ -65,7 +65,7 @@
 }
 
 @article{Sullivan:2019,
-  author = {Sullivan, Ian AND DeHaven, Alexander AND Mellor, David},
+  author = {Sullivan, Ian and DeHaven, Alexander and Mellor, David},
   title = {Open and Reproducible Research on Open Science Framework},
   journal = {Current protocols},
   year = {2019},
@@ -77,7 +77,7 @@
 }
 
 @article{Wilson:2017,
-  author = {Wilson, Greg AND Bryan, Jennifer AND Cranston, Karen AND Kitzes, Justin AND Nederbragt, Lex AND Teal, Tracy K.},
+  author = {Wilson, Greg and Bryan, Jennifer and Cranston, Karen and Kitzes, Justin and Nederbragt, Lex and Teal, Tracy K.},
   journal = {PLOS Computational Biology},
   publisher = {Public Library of Science},
   title = {Good enough practices in scientific computing},
