+++ inst/paper.bib
@@ -30,7 +30,8 @@
   volume = {8},
   number = {1},
   pages = {7},
-  month = feb
+  month = feb,
+  doi = {10.1186/1751-0473-8-7}
 }
 
 @article{Sandve:2013,
