### 2020-02-05 981f78c
Exclude license from build
