### 2020-02-05 8c9f411
Update CRAN comments
