### 2020-02-06 dd49b49
Add Makefile to simplify release prep
