+++ .github/CONTRIBUTING.md
@@ -40,25 +40,40 @@ the GitHub web interface, so long as the changes are made in the _source_ file.
 
 Before you make a substantial pull request, you should always file an issue and
 make sure someone from the team agrees that it’s a problem. If you’ve found a
-bug, create an associated issue and illustrate the bug with a minimal 
+bug, create an associated issue and illustrate the bug with a minimal
 [reprex](https://www.tidyverse.org/help/#reprex).
 
 ### Pull request process
 
-*  We recommend that you create a Git branch for each pull request (PR).  
+*  We recommend that you create a Git branch for each pull request (PR).
 *  Look at the Travis and AppVeyor build status before and after making changes.
 The `README` should contain badges for any continuous integration services used
-by the package.  
+by the package.
 *  New code should follow the tidyverse [style guide](http://style.tidyverse.org).
 You can use the [styler](https://CRAN.R-project.org/package=styler) package to
-apply these styles, but please don't restyle code that has nothing to do with 
-your PR.  
+apply these styles, but please don't restyle code that has nothing to do with
+your PR.
 *  We use [roxygen2](https://cran.r-project.org/package=roxygen2), with
-[Markdown syntax](https://cran.r-project.org/web/packages/roxygen2/vignettes/markdown.html), 
-for documentation.  
+[Markdown syntax](https://cran.r-project.org/web/packages/roxygen2/vignettes/markdown.html),
+for documentation.
 *  We use [testthat](https://cran.r-project.org/package=testthat). Contributions
-with test cases included are easier to accept.  
+with test cases included are easier to accept.
 *  For user-facing changes, add a bullet to the top of `NEWS.md` below the
 current development version header describing the changes made followed by your
 GitHub username, and links to relevant issue(s)/PR(s).
 
+### Using the included `Makefile`
+
+Run `make docs` to:
+* rebuild `README.md` from `README.Rmd`
+* regenerate the precomputed vignette `getting_started.Rmd` from `getting_started.Rmd.orig`
+* rebuild the vignettes
+* rebuild package documentation
+
+Run `make` to
+* perform all of the documentation steps noted above
+* build the package
+* check the package as CRAN but without running tests (this is temporary until mock tests are implemented)
+
+Run `make test` to:
+* run unit tests
