### 2020-02-06 9f560a1
Update links to point to ropensci docs
