+++ README.md
@@ -1,36 +1,30 @@
+---
+output: github_document
+---
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-# osfr <a href="https://ropensci.github.io/osfr/"><img src="man/figures/logo.png" align="right" height="139" /></a>
 
-[![Build
-Status](https://travis-ci.com/ropensci/osfr.svg)](https://travis-ci.com/ropensci/osfr)
-[![AppVeyor build
-status](https://ci.appveyor.com/api/projects/status/github/ropensci/osfr?branch=master&svg=true)](https://ci.appveyor.com/project/aaronwolen/osfr)
-[![Coverage
-status](https://codecov.io/gh/ropensci/osfr/branch/master/graph/badge.svg)](https://codecov.io/github/ropensci/osfr?branch=master)
+
+# osfr <a href="https://docs.ropensci.org/osfr"><img src="man/figures/logo.png" align="right" height="139" /></a>
+
+[![Build Status](https://travis-ci.com/ropensci/osfr.svg)](https://travis-ci.com/ropensci/osfr)
+[![AppVeyor build status](https://ci.appveyor.com/api/projects/status/github/ropensci/osfr?branch=master&svg=true)](https://ci.appveyor.com/project/aaronwolen/osfr)
+[![Coverage status](https://codecov.io/gh/ropensci/osfr/branch/master/graph/badge.svg)](https://codecov.io/github/ropensci/osfr?branch=master)
 [![](https://badges.ropensci.org/279_status.svg)](https://github.com/ropensci/software-review/issues/279)
 [![DOI](https://zenodo.org/badge/42329785.svg)](https://zenodo.org/badge/latestdoi/42329785)
 
 ## Overview
 
-osfr provides a suite of functions for interacting with the Open Science
-Framework ([OSF](https://osf.io "Open Science Framework")).
+osfr provides a suite of functions for interacting with the Open Science Framework ([OSF][osf]).
 
 **What is OSF?**
 
-*OSF is a free and [open
-source](https://github.com/CenterForOpenScience/osf.io "OSF's GitHub Repository")
-project management repository designed to support researchers across
-their entire project lifecycle. The service includes unlimited cloud
-storage and file version history, providing a centralized location for
-all your research materials that can be kept private, shared with select
-collaborators, or made publicly available with citable DOIs.*
+*OSF is a free and [open source][osf-gh] project management repository designed to support researchers across their entire project lifecycle. The service includes unlimited cloud storage and file version history, providing a centralized location for all your research materials that can be kept private, shared with select collaborators, or made publicly available with citable DOIs.*
 
 ## Installation
 
-This package is currently under development and is not available on
-CRAN. You can install the current version with the *remotes* package:
+This package is currently under development and is not available on CRAN. You can install the current version with the *remotes* package:
 
 ``` r
 # install.packages("remotes")
@@ -39,78 +33,60 @@ remotes::install_github("ropensci/osfr")
 
 ## Usage Examples
 
-*Note: You need to [setup an OSF personal access token
-(PAT)](https://ropensci.github.io/osfr/articles/auth) to use osfr to
-manage projects or upload files.*
+*Note: You need to [setup an OSF personal access token (PAT)][auth] to use osfr to manage projects or upload files.*
 
 ### Accessing Open Research Materials
 
-Many researchers use OSF to archive and share their work. You can use
-osfr to explore publicly accessible projects and download the associated
-files—all you need to get started is the project’s URL or GUID (global
-unique identifier).
[…]
