+++ vignettes/getting_started.Rmd
@@ -1,7 +1,7 @@
 ---
 title: "Getting Started with osfr"
-date: "2020-01-23"
-output: 
+date: "2020-02-06"
+output:
   rmarkdown::html_vignette:
     toc: TRUE
 vignette: >
@@ -12,7 +12,7 @@ vignette: >
 
 
 
-This vignette provides a quick tour of the *osfr* package. 
+This vignette provides a quick tour of the *osfr* package.
 
 
 ```r
@@ -21,9 +21,9 @@ library(osfr)
 
 
 
-## What is OSF? 
+## What is OSF?
 
-[OSF][osf] is a free and open source web application that provides a space for researchers to collaboratively store, manage, and share their research materials (e.g. data, code, protocols). 
+[OSF][osf] is a free and open source web application that provides a space for researchers to collaboratively store, manage, and share their research materials (e.g. data, code, protocols).
 
 Most work on OSF is organized around ***projects***, which include a cloud-based storage bucket where files can be stored and organized into directories. Note there is no storage limit on the size of projects but individual files must be < 5Gb. Projects can be kept private, shared with a specific group of collaborators, or made publicly available with citable DOIs so you can get credit for their work.
 
@@ -36,7 +36,7 @@ Here are a few relevant topics:
 
 ## Accessing OSF projects
 
-Let's check out an example project containing materials for an analysis of the 2012 American National Election Survey (ANES). You can access the OSF project in your browser by navigating to its URL: <https://osf.io/jgyxm/>. 
+Let's check out an example project containing materials for an analysis of the 2012 American National Election Survey (ANES). You can access the OSF project in your browser by navigating to its URL: <https://osf.io/jgyxm/>.
 
 Let's load this project into R with `osfr::osf_retrieve_node()`:
 
@@ -50,7 +50,7 @@ anes_project
 #> 1 Political identification and gender jgyxm <named list [3]>
 ```
 
-This returns an `osf_tbl` object, which is the `data.frame`-like class *osfr* uses to represent items retrieved from OSF. You can now use `anes_project` to perform a variety of project related tasks by passing it to different osfr functions. 
+This returns an `osf_tbl` object, which is the `data.frame`-like class *osfr* uses to represent items retrieved from OSF. You can now use `anes_project` to perform a variety of project related tasks by passing it to different osfr functions.
 
 ### Downloading files
 
@@ -81,19 +81,19 @@ Because `osf_tbl`s are just specialized `data.frame`s, we could also `subset()` 
 
 *__Note:__ If an `osf_tbl` with multiple entities is passed to an non-vectorized osfr function like `osf_open()`, the default behavior is to use the entity in the first row and warn that all other entities are ignored.*
 
-We can also download local copies of these files by passing `anes_files` to `osf_download()`. 
+We can also download local copies of these files by passing `anes_files` to `osf_download()`.
 
 
 ```r
 osf_download(anes_files)
 #> # A tibble: 5 x 4
-#>   name                 id                       local_path             meta            
-#>   <chr>                <chr>                    <chr>                  <list>          
-#> 1 cleaning.R           5e20d22bedceab002d82e0f1 ./cleaning.R           <named list [3]>
-#> 2 Questionnaire.docx   5e20d22bedceab002b82dc3f ./Questionnaire.docx   <named list [3]>
-#> 3 raw_data.csv         5e20d22c675e0e00096b4de8 ./raw_data.csv         <named list [3]>
[…]
