### 2020-02-06 ecd2c45
Add JOSS badge
