+++ README.md
@@ -12,6 +12,7 @@ output: github_document
 [![AppVeyor build status](https://ci.appveyor.com/api/projects/status/github/ropensci/osfr?branch=master&svg=true)](https://ci.appveyor.com/project/aaronwolen/osfr)
 [![Coverage status](https://codecov.io/gh/ropensci/osfr/branch/master/graph/badge.svg)](https://codecov.io/github/ropensci/osfr?branch=master)
 [![](https://badges.ropensci.org/279_status.svg)](https://github.com/ropensci/software-review/issues/279)
+[![](https://joss.theoj.org/papers/d5398fc36ea92794a20914143d3fcdc4/status.svg)](https://joss.theoj.org/papers/d5398fc36ea92794a20914143d3fcdc4)
 [![DOI](https://zenodo.org/badge/42329785.svg)](https://zenodo.org/badge/latestdoi/42329785)
 
 ## Overview
@@ -50,8 +51,8 @@ library(osfr)
 cr_project <- osf_retrieve_node("e81xl")
 cr_project
 #> # A tibble: 1 x 3
-#>   name                                    id    meta
-#>   <chr>                                   <chr> <list>
+#>   name                                    id    meta            
+#>   <chr>                                   <chr> <list>          
 #> 1 Reproducibility Project: Cancer Biology e81xl <named list [3]>
 ```
 
@@ -61,8 +62,8 @@ This returns an `osf_tbl` object with a single row representing the retrieved pr
 ```r
 osf_ls_files(cr_project)
 #> # A tibble: 4 x 3
-#>   name                                     id                     meta
-#>   <chr>                                    <chr>                  <list>
+#>   name                                     id                     meta          
+#>   <chr>                                    <chr>                  <list>        
 #> 1 Adjustment of 50 studies to 37 studies.… 565602398c5e4a3877d72… <named list […
 #> 2 papers_and_keywords.xlsx                 553e671b8c5e4a219919e… <named list […
 #> 3 Full_dataset_of_papers_formatted.xls     553e671b8c5e4a219919e… <named list […
@@ -77,8 +78,8 @@ This project contains 2 ***components***: *Replication Studies* and *Data collec
 ```r
 osf_ls_nodes(cr_project)
 #> # A tibble: 2 x 3
-#>   name                                      id    meta
-#>   <chr>                                     <chr> <list>
+#>   name                                      id    meta            
+#>   <chr>                                     <chr> <list>          
 #> 1 Replication Studies                       p7ayb <named list [3]>
 #> 2 Data collection and publishing guidelines a5imq <named list [3]>
 ```
@@ -95,8 +96,8 @@ cr_project %>%
   osf_ls_nodes(pattern = "Study 19") %>%
   osf_ls_files()
 #> # A tibble: 6 x 3
-#>   name                                    id                     meta
-#>   <chr>                                   <chr>                  <list>
+#>   name                                    id                     meta           
+#>   <chr>                                   <chr>                  <list>         
 #> 1 Replication_Study_19.docx               57c9e8ed594d9001e7a24… <named list [3…
 #> 2 Replication_Study_19.Rmd                578e2b23594d9001f4816… <named list [3…
 #> 3 Replication_Study_19_track_changes.docx 581a27b76c613b0223322… <named list [3…
@@ -112,8 +113,8 @@ We could continue this pattern of exploration and even download local copies of 
 osf_retrieve_file("https://osf.io/btgx3/") %>%
   osf_download()
 #> # A tibble: 1 x 4
-#>   name                id                    local_path            meta
-#>   <chr>               <chr>                 <chr>                 <list>
+#>   name                id                    local_path            meta          
[…]
