### 2020-02-06 5c04bea
Update cran comments
