+++ cran-comments.md
@@ -14,9 +14,10 @@
 
 This is a resubmission. In this version I have:
 
-- replaced links in README.md with fully specified URLs
+- replaced github links in README.md with fully specified URLs
 - updated the TITLE and DESCRIPTION to place 'osfr', 'OSF', and 'Open Science Framework' in quotes
-- removed the full license builds
+- removed the full license from builds
+- fixed invalid README URLs, which resulted from migrating to rOpenSci's docs server
 
 ## Reverse dependencies
 
