### 2020-02-06 458b037
Render README using devtools
Properly removes the yaml metadata from the markdown file and inserts
all reference links inline. [skip ci]
