### 2020-02-06 aaa6844
Add CITATION
[skip ci]
