### 2020-02-17 89d3d69
Adjust remote path when uploading to a directory
Fixes #121
