### 2020-02-17 1bfbbbd
Switch to MIT license and update cran comments
[skip-ci]
