+++ DESCRIPTION
@@ -30,7 +30,7 @@ Imports:
     fs (>= 1.3.0),
     memoise,
     httr
-License: Apache License (>= 2.0) | file LICENSE
+License: MIT + file LICENSE
 URL:
     https://docs.ropensci.org/osfr,
     https://github.com/ropensci/osfr
