+++ NEWS.md
@@ -1,3 +1,5 @@
+# osfr (development version)
+
 # osfr 0.2.8
 
 * Initial CRAN release
