+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: osfr
 Title: Interface to the 'Open Science Framework' ('OSF')
-Version: 0.2.8
+Version: 0.2.8.9000
 Authors@R: c(
     person("Aaron", "Wolen",, "aaron@wolen.com", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0003-2542-2202")),
