### 2020-02-18 0830897
Add CRAN badge and install instructions
