+++ README.Rmd
@@ -17,6 +17,7 @@ Sys.unsetenv(c("OSF_SERVER", "OSF_PAT"))
 
 # osfr <a href="https://docs.ropensci.org/osfr"><img src="man/figures/logo.png" align="right" height="139" /></a>
 
+[![CRAN status](https://www.r-pkg.org/badges/version/osfr)](https://CRAN.R-project.org/package=osfr)
 [![Build Status](https://travis-ci.com/ropensci/osfr.svg)](https://travis-ci.com/ropensci/osfr)
 [![AppVeyor build status](https://ci.appveyor.com/api/projects/status/github/ropensci/osfr?branch=master&svg=true)](https://ci.appveyor.com/project/aaronwolen/osfr)
 [![Coverage status](https://codecov.io/gh/ropensci/osfr/branch/master/graph/badge.svg)](https://codecov.io/github/ropensci/osfr?branch=master)
@@ -34,7 +35,13 @@ osfr provides a suite of functions for interacting with the Open Science Framewo
 
 ## Installation
 
-This package is currently under development and is not available on CRAN. You can install the current version with the *remotes* package:
+You can install the current release of osfr from CRAN (*recommended*):
+
+```r
+install.packages("osfr")
+```
+
+Or the development version from GitHub with the *remotes* package:
 
 ``` r
 # install.packages("remotes")
