+++ tests/testthat/cassettes/directories/create-dir1.yml
@@ -1,14 +1,14 @@
 http_interactions:
 - request:
     method: get
-    uri: https://api.test.osf.io/v2/nodes/mdfb8/files/osfstorage/?filter%5Bkind%5D=folder&filter%5Bname%5D=dir1&page=1
+    uri: https://api.test.osf.io/v2/nodes/drwby/files/osfstorage/?filter%5Bkind%5D=folder&filter%5Bname%5D=dir1&page=1
     body:
       encoding: ''
       string: ''
     headers:
       Accept-Encoding: gzip, deflate
       Accept: application/json, text/xml, application/xml, */*
-      User-Agent: osfr v0.2.8
+      User-Agent: osfr v0.2.8.9000
       Authorization: Bearer <totallyrealpat>
       Accept-Header: application/vnd.api+json;version=2.8
   response:
@@ -22,7 +22,7 @@ http_interactions:
       cache-control: no-cache, no-store, max-age=0, must-revalidate
       content-length: '128'
       content-type: application/vnd.api+json; charset=utf-8
-      date: Wed, 29 Jan 2020 23:27:21 GMT
+      date: Wed, 19 Feb 2020 18:32:56 GMT
       expires: Mon, 01 Jan 1990 00:00:00 GMT
       pragma: no-cache
       server: nginx
@@ -36,18 +36,18 @@ http_interactions:
       encoding: UTF-8
       file: no
       string: '{"data":[],"links":{"first":null,"last":null,"prev":null,"next":null,"meta":{"total":0,"per_page":10}},"meta":{"version":"2.0"}}'
-  recorded_at: 2020-01-29 23:27:23 GMT
-  recorded_with: vcr/0.4.1.93.9000, webmockr/0.5.1.96
+  recorded_at: 2020-02-19 18:32:57 GMT
+  recorded_with: vcr/0.4.1.97, webmockr/0.5.1.96
 - request:
     method: put
-    uri: https://files.us.test.osf.io/v1/resources/mdfb8/providers/osfstorage/?kind=folder&name=dir1
+    uri: https://files.us.test.osf.io/v1/resources/drwby/providers/osfstorage/?kind=folder&name=dir1
     body:
       encoding: ''
       string: ''
     headers:
       Accept-Encoding: gzip, deflate
       Accept: application/json, text/xml, application/xml, */*
-      User-Agent: osfr v0.2.8
+      User-Agent: osfr v0.2.8.9000
       Authorization: Bearer <totallyrealpat>
   response:
     status:
@@ -57,10 +57,10 @@ http_interactions:
     headers:
       status: HTTP/1.1 201 Created
       server: nginx
-      date: Wed, 29 Jan 2020 23:27:22 GMT
+      date: Wed, 19 Feb 2020 18:32:56 GMT
       content-type: application/json; charset=UTF-8
       content-length: '821'
-      x-waterbutler-request-id: e657fb65-cec4-4e28-a993-3812fc1bcfd9
+      x-waterbutler-request-id: 0d612924-aac6-42fe-ba59-13f8a3cac5b8
       cache-control: no-cache, no-store, max-age=0, must-revalidate
       expires: Mon, 01 Jan 1990 00:00:00 GMT
       pragma: no-cache
@@ -69,27 +69,25 @@ http_interactions:
     body:
       encoding: UTF-8
       file: no
-      string: '{"data": {"type": "files", "id": "osfstorage/5e3214da342fc900096700b6/",
-        "links": {"new_folder": "https://files.us.test.osf.io/v1/resources/mdfb8/providers/osfstorage/5e3214da342fc900096700b6/?kind=folder",
[…]
