+++ tests/testthat/cassettes/directories/list-within-subdir-dir1.yml
@@ -1,14 +1,14 @@
 http_interactions:
 - request:
     method: get
-    uri: https://api.test.osf.io/v2/nodes/mdfb8/files/osfstorage/?filter%5Bkind%5D=folder&filter%5Bname%5D=dir1&page=1
+    uri: https://api.test.osf.io/v2/nodes/drwby/files/osfstorage/?filter%5Bkind%5D=folder&filter%5Bname%5D=dir1&page=1
     body:
       encoding: ''
       string: ''
     headers:
       Accept-Encoding: gzip, deflate
       Accept: application/json, text/xml, application/xml, */*
-      User-Agent: osfr v0.2.8
+      User-Agent: osfr v0.2.8.9000
       Authorization: Bearer <totallyrealpat>
       Accept-Header: application/vnd.api+json;version=2.8
   response:
@@ -22,7 +22,7 @@ http_interactions:
       cache-control: no-cache, no-store, max-age=0, must-revalidate
       content-encoding: gzip
       content-type: application/vnd.api+json; charset=utf-8
-      date: Wed, 29 Jan 2020 23:27:44 GMT
+      date: Wed, 19 Feb 2020 18:32:59 GMT
       expires: Mon, 01 Jan 1990 00:00:00 GMT
       pragma: no-cache
       server: nginx
@@ -36,19 +36,19 @@ http_interactions:
     body:
       encoding: UTF-8
       file: no
-      string: '{"data":[{"id":"5e3214da342fc900096700b6","type":"files","attributes":{"guid":null,"checkout":null,"name":"dir1","kind":"folder","path":"/5e3214da342fc900096700b6/","size":null,"provider":"osfstorage","materialized_path":"/dir1/","last_touched":null,"date_modified":null,"date_created":null,"extra":{"hashes":{"md5":null,"sha256":null}},"tags":[],"current_user_can_comment":true,"current_version":1},"relationships":{"parent_folder":{"links":{"related":{"href":"https://api.test.osf.io/v2/files/5e32148b58cd6b00090d9a7a/","meta":{}}},"data":{"id":"5e32148b58cd6b00090d9a7a","type":"files"}},"files":{"links":{"related":{"href":"https://api.test.osf.io/v2/nodes/mdfb8/files/osfstorage/5e3214da342fc900096700b6/","meta":{}}}},"node":{"links":{"related":{"href":"https://api.test.osf.io/v2/nodes/mdfb8/","meta":{}}},"data":{"id":"mdfb8","type":"nodes"}},"target":{"links":{"related":{"href":"https://api.test.osf.io/v2/nodes/mdfb8/","meta":{"type":"node"}}},"data":{"type":"node","id":"mdfb8"}}},"links":{"info":"https://api.test.osf.io/v2/files/5e3214da342fc900096700b6/","move":"https://files.us.test.osf.io/v1/resources/mdfb8/providers/osfstorage/5e3214da342fc900096700b6/","upload":"https://files.us.test.osf.io/v1/resources/mdfb8/providers/osfstorage/5e3214da342fc900096700b6/","delete":"https://files.us.test.osf.io/v1/resources/mdfb8/providers/osfstorage/5e3214da342fc900096700b6/","new_folder":"https://files.us.test.osf.io/v1/resources/mdfb8/providers/osfstorage/5e3214da342fc900096700b6/?kind=folder","self":"https://api.test.osf.io/v2/files/5e3214da342fc900096700b6/"}}],"links":{"first":null,"last":null,"prev":null,"next":null,"meta":{"total":1,"per_page":10}},"meta":{"version":"2.0"}}'
[…]
