+++ tests/testthat/cassettes/directories/list-empty-p1.yml
@@ -1,14 +1,14 @@
 http_interactions:
 - request:
     method: get
-    uri: https://api.test.osf.io/v2/nodes/mdfb8/files/osfstorage/?page=1
+    uri: https://api.test.osf.io/v2/nodes/drwby/files/osfstorage/?page=1
     body:
       encoding: ''
       string: ''
     headers:
       Accept-Encoding: gzip, deflate
       Accept: application/json, text/xml, application/xml, */*
-      User-Agent: osfr v0.2.8
+      User-Agent: osfr v0.2.8.9000
       Authorization: Bearer <totallyrealpat>
       Accept-Header: application/vnd.api+json;version=2.8
   response:
@@ -22,7 +22,7 @@ http_interactions:
       cache-control: no-cache, no-store, max-age=0, must-revalidate
       content-length: '128'
       content-type: application/vnd.api+json; charset=utf-8
-      date: Wed, 29 Jan 2020 23:27:01 GMT
+      date: Wed, 19 Feb 2020 18:32:55 GMT
       expires: Mon, 01 Jan 1990 00:00:00 GMT
       pragma: no-cache
       server: nginx
@@ -36,5 +36,5 @@ http_interactions:
       encoding: UTF-8
       file: no
       string: '{"data":[],"links":{"first":null,"last":null,"prev":null,"next":null,"meta":{"total":0,"per_page":10}},"meta":{"version":"2.0"}}'
-  recorded_at: 2020-01-29 23:27:01 GMT
-  recorded_with: vcr/0.4.1.93.9000, webmockr/0.5.1.96
+  recorded_at: 2020-02-19 18:32:55 GMT
+  recorded_with: vcr/0.4.1.97, webmockr/0.5.1.96

[…]
